Movies (data)

The data concerns a sample of 1000 Movies which were on screens in France and come of their characteristics.

data(Movies)

Format

A data frame with 1000 observations and the following 7 variables:

Budget

numeric vector of movie budgets

Genre

is a factor with 9 levels

Country

is a factor with 4 level. Country of origin of the movie.

ArtHouse

is a factor with levels No, Yes. Whether the movie had the "Art House" label.

Festival

is a factor with levels No, Yes. Whether the movie was selected in Cannes, Berlin or Venise film festivals.

Critics

numeric vector of average ratings from intellectual criticism.

BoxOffice

numeric vector of number of admissions.

Examples

data(Movies)
str(Movies)
#> 'data.frame':	1000 obs. of  7 variables:
#>  $ Budget   : num  3.10e+07 4.88e+06 3.50e+06 1.63e+08 2.17e+07 ...
#>  $ Genre    : Factor w/ 9 levels "Action","Animation",..: 1 5 7 1 7 5 1 7 5 7 ...
#>  $ Country  : Factor w/ 4 levels "Europe","France",..: 4 2 2 1 2 2 4 4 2 4 ...
#>  $ ArtHouse : Factor w/ 2 levels "No","Yes": 1 1 2 1 2 1 1 1 1 1 ...
#>  $ Festival : Factor w/ 2 levels "No","Yes": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ Critics  : num  3 1 3.75 3.75 3.6 2.75 1 1 1 3 ...
#>  $ BoxOffice: num  1013509 24241 39376 6996996 493416 ...