Taste (data)

The data concerns tastes for music and movies of a set of 2000 individuals. It contains 5 variables of likes for music genres (french pop, rap, rock, jazz and classical), 6 variables of likes for movie genres (comedy, crime, animation, science fiction, love, musical) and 2 additional variables (gender and age).

data(Taste)

Format

A data frame with 500 observations and the following 13 variables:

FrenchPop

is a factor with levels No, Yes, NA

Rap

is a factor with levels No, Yes, NA

Rock

is a factor with levels No, Yes, NA

Jazz

is a factor with levels No, Yes, NA

Classical

is a factor with levels No, Yes, NA

Comedy

is a factor with levels No, Yes, NA

Crime

is a factor with levels No, Yes, NA

Animation

is a factor with levels No, Yes, NA

SciFi

is a factor with levels No, Yes, NA

Love

is a factor with levels No, Yes, NA

Musical

is a factor with levels No, Yes, NA

Gender

is a factor with levels Men, Women

Age

is a factor with levels 15-24, 25-49, 50+

Educ

is a factor with levels none, low, medium, high

Details

'NA' stands for 'not available'

Examples

data(Taste)
str(Taste)

#> 'data.frame':	2000 obs. of  14 variables:
#>  $ FrenchPop: Factor w/ 3 levels "No","Yes","NA": 2 1 2 1 2 1 1 1 1 2 ...
#>  $ Rap      : Factor w/ 3 levels "No","Yes","NA": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ Rock     : Factor w/ 3 levels "No","Yes","NA": 1 1 2 1 1 2 1 1 2 1 ...
#>  $ Jazz     : Factor w/ 3 levels "No","Yes","NA": 1 2 1 1 1 1 1 1 1 1 ...
#>  $ Classical: Factor w/ 3 levels "No","Yes","NA": 1 2 1 2 1 1 1 1 1 1 ...
#>  $ Comedy   : Factor w/ 3 levels "No","Yes","NA": 1 2 1 1 1 1 2 2 2 2 ...
#>  $ Crime    : Factor w/ 3 levels "No","Yes","NA": 1 1 1 1 2 1 1 1 1 1 ...
#>  $ Animation: Factor w/ 3 levels "No","Yes","NA": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ SciFi    : Factor w/ 3 levels "No","Yes","NA": 2 1 1 1 1 2 1 1 1 1 ...
#>  $ Love     : Factor w/ 3 levels "No","Yes","NA": 1 1 2 1 1 1 1 1 1 1 ...
#>  $ Musical  : Factor w/ 3 levels "No","Yes","NA": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ Gender   : Factor w/ 2 levels "Men","Women": 1 1 2 1 2 2 2 2 1 1 ...
#>  $ Age      : Factor w/ 3 levels "15-24","25-49",..: 2 3 2 3 2 2 2 2 1 3 ...
#>  $ Educ     : Factor w/ 4 levels "None","Low","Medium",..: 3 4 3 4 2 1 3 2 2 2 ...