Cross-tabulation and measures of association between two categorical variables

Cross-tabulation and measures of association between two categorical variables

assoc.twocat(x,y,weights=rep.int(1,length(x)),na_value=NULL,nperm=NULL,distrib="asympt")

Arguments

x
the first categorical variable (must be a factor)

y
the second categorical variable (must be a factor)

weights
an optional numeric vector of weights (by default, a vector of 1 for uniform weights)

na_value
character. Name of the level for NA category. If NULL (default), NA values are ignored.

nperm
numeric. Number of permutations for the permutation test of independence. If NULL (default), no permutation test is performed. Default is 1000.

distrib
the null distribution of permutation test of independence can be approximated by its asymptotic distribution ("asympt", default) or via Monte Carlo resampling ("approx".

Value

A list with the following elements :

freq

cross-tabulation

prop

percentages

rprop

row percentages

cprop

column percentages

expected

expected values

chi.squared

chi-squared value

cramer.v

Cramer's V between the two variables

permutation.pvalue

p-value from a permutation (so non-parametric) test of independence

pearson.residuals

the table of Pearson residuals, i.e. (observed - expected) / sqrt(expected).

phi

the table of the phi coefficients for each pair of levels

phi.perm.pval

the table of permutation p-values for each pair of levels

local.pem

the table of local Percentages of Maximum Deviation from Independences

global.pem

value of the global Percentage of Maximum Deviation from Independence

gather

a data frame gathering informations, with one row per cell of the cross-tabulation

References

Rakotomalala R., 'Comprendre la taille d'effet (effect size)', http://eric.univ-lyon2.fr/~ricco/cours/slides/effect_size.pdf

Author

Nicolas Robette

See also

Examples

data(Music)
assoc.twocat(Music$Jazz,Music$Age,nperm=100)

#> $freq
#>     15-24 25-49 50+ Sum
#> No     63   172 172 407
#> Yes     8    28  50  86
#> NA      2     2   3   7
#> Sum    73   202 225 500
#> 
#> $prop
#>     15-24 25-49   50+   Sum
#> No   12.6  34.4  34.4  81.4
#> Yes   1.6   5.6  10.0  17.2
#> NA    0.4   0.4   0.6   1.4
#> Sum  14.6  40.4  45.0 100.0
#> 
#> $rprop
#>     15-24 25-49  50+ Sum
#> No   15.5  42.3 42.3 100
#> Yes   9.3  32.6 58.1 100
#> NA   28.6  28.6 42.9 100
#> Sum  14.6  40.4 45.0 100
#> 
#> $cprop
#>     15-24 25-49   50+   Sum
#> No   86.3  85.1  76.4  81.4
#> Yes  11.0  13.9  22.2  17.2
#> NA    2.7   1.0   1.3   1.4
#> Sum 100.0 100.0 100.0 100.0
#> 
#> $expected
#>       15-24   25-49     50+
#> No   59.422 164.428 183.150
#> Yes  12.556  34.744  38.700
#> NA    1.022   2.828   3.150
#> 
#> $chi.squared
#> [1] 8.690099
#> 
#> $cramer.v
#> [1] 0.0932207
#> 
#> $permutation.pvalue
#> [1] 0.03099653
#> 
#> $pearson.residuals
#>           15-24       25-49         50+
#> No   0.46415892  0.59050383 -0.82389412
#> Yes -1.28575453 -1.14413601  1.81644796
#> NA   0.96741631 -0.49236892 -0.08451543
#> 
#> $phi
#>            15-24        25-49          50+
#> No   0.052083036  0.079315509 -0.115198968
#> Yes -0.068379978 -0.072837386  0.120376426
#> NA   0.047147754 -0.028723942 -0.005132526
#> 
#> $phi.perm.pval
#>           15-24       25-49         50+
#> No  0.082157126 0.059619026 0.005458039
#> Yes 0.043259335 0.073358659 0.004336973
#> NA  0.159489636 0.263636387 0.469698175
#> 
#> $local.pem
#>     15-24 25-49   50+
#> No   26.4  20.2 -21.8
#> Yes -36.3 -19.4  23.9
#> NA   16.4 -29.3  -4.8
#> 
#> $global.pem
#> [1] 24
#> 
#> $gather
#>   Var1  Var2 Freq  prop      rprop      cprop expected std.residuals
#> 1   No 15-24   63 0.126 0.15479115 0.86301370   59.422    0.46415892
#> 2  Yes 15-24    8 0.016 0.09302326 0.10958904   12.556   -1.28575453
#> 3   NA 15-24    2 0.004 0.28571429 0.02739726    1.022    0.96741631
#> 4   No 25-49  172 0.344 0.42260442 0.85148515  164.428    0.59050383
#> 5  Yes 25-49   28 0.056 0.32558140 0.13861386   34.744   -1.14413601
#> 6   NA 25-49    2 0.004 0.28571429 0.00990099    2.828   -0.49236892
#> 7   No   50+  172 0.344 0.42260442 0.76444444  183.150   -0.82389412
#> 8  Yes   50+   50 0.100 0.58139535 0.22222222   38.700    1.81644796
#> 9   NA   50+    3 0.006 0.42857143 0.01333333    3.150   -0.08451543
#>            phi   perm.pval local.pem
#> 1  0.052083036 0.082157126      26.4
#> 2 -0.068379978 0.043259335     -36.3
#> 3  0.047147754 0.159489636      16.4
#> 4  0.079315509 0.059619026      20.2
#> 5 -0.072837386 0.073358659     -19.4
#> 6 -0.028723942 0.263636387     -29.3
#> 7 -0.115198968 0.005458039     -21.8
#> 8  0.120376426 0.004336973      23.9
#> 9 -0.005132526 0.469698175      -4.8
#>