Bootstrap validation for MCA (supplementary variables)

Bootstrap validation of MCA, through the computation of the coordinates of supplementary variables for bootstrap replications of the data.

bootvalid_supvars(resmca, vars = NULL, axes = c(1,2), K = 30)

Arguments

resmca

object of class speMCA.

vars

A data frame of categorical supplementary variables. All these variables should be factors.

axes

numeric vector of length 2, specifying the components (axes) to plot. Default is c(1,2).

K

integer. Number of bootstrap replications (default is 30).

Details

The bootstrap technique is used here as an internal (and non-parametric) validation procedure of the results of a multiple correspondence analysis. For supplementary variables, only partial bootstrap is possible. The partial bootstrap does not compute new MCAs: it projects bootstrap replications of the initial data as supplementary elements of the MCA. See references for more details.

Value

A data frame with the following elements :

varcat

Name of the active category

K

Index of the bootstrap replication

dim.x

Bootstrap coordinate on the first selected axis

dim.y

Bootstrap coordinate on the second selected axis

References

Lebart L. (2006). "Validation Techniques in Multiple Correspondence Analysis". In M. Greenacre et J. Blasius (eds), Multiple Correspondence Analysis and related techniques, Chapman and Hall/CRC, p.179-196.

Lebart L. (2007). "Which bootstrap for principal axes methods?". In P. Brito et al. (eds), Selected Contributions in Data Analysis and Classification, Springer, p.581-588.

Author

Nicolas Robette

See also

Examples

data(Taste)
resmca <- speMCA(Taste[,1:11], excl=c(3,6,9,12,15,18,21,24,27,30,33))
supvars <- Taste[,c("Gender", "Age", "Educ")]
bootvalid_supvars(resmca, supvars, K = 5)
#>          varcat K        dim.1         dim.2
#> 3     Age.15-24 1  0.431065155  0.6328054188
#> 12    Age.15-24 2  0.422531632  0.7178661918
#> 21    Age.15-24 3  0.436520483  0.6313169931
#> 30    Age.15-24 4  0.442775684  0.7723383955
#> 39    Age.15-24 5  0.531872015  0.7276109723
#> 4     Age.25-49 1  0.006033626  0.1419212042
#> 13    Age.25-49 2  0.033803065  0.2096337124
#> 22    Age.25-49 3  0.056336272  0.1377075680
#> 31    Age.25-49 4  0.018100178  0.1820962015
#> 40    Age.25-49 5  0.036519910  0.1153170702
#> 5       Age.50+ 1 -0.136276086 -0.3738340777
#> 14      Age.50+ 2 -0.137922957 -0.3815194151
#> 23      Age.50+ 3 -0.149873754 -0.4147842206
#> 32      Age.50+ 4 -0.179019606 -0.3930194843
#> 41      Age.50+ 5 -0.139725851 -0.3577117226
#> 9     Educ.High 1 -0.338665720  0.2332623077
#> 18    Educ.High 2 -0.328155276  0.2976238844
#> 27    Educ.High 3 -0.368833313  0.2018978688
#> 36    Educ.High 4 -0.374388336  0.2340005882
#> 45    Educ.High 5 -0.369801266  0.2450609575
#> 7      Educ.Low 1  0.152437759  0.0064311099
#> 16     Educ.Low 2  0.122797881  0.0767468862
#> 25     Educ.Low 3  0.117714743 -0.0265903121
#> 34     Educ.Low 4  0.127148882  0.0633443936
#> 43     Educ.Low 5  0.169560691  0.0319780866
#> 8   Educ.Medium 1 -0.077259177  0.0006705646
#> 17  Educ.Medium 2 -0.112996758  0.0347423902
#> 26  Educ.Medium 3 -0.035567759  0.0108634393
#> 35  Educ.Medium 4 -0.158757904  0.0921398401
#> 44  Educ.Medium 5 -0.041482284  0.0729135654
#> 6     Educ.None 1  0.231446885 -0.2640275300
#> 15    Educ.None 2  0.255851082 -0.3019425240
#> 24    Educ.None 3  0.295512999 -0.2628419583
#> 33    Educ.None 4  0.254213141 -0.2878788966
#> 42    Educ.None 5  0.291660044 -0.3075167431
#> 1    Gender.Men 1 -0.056746891  0.2726170594
#> 10   Gender.Men 2  0.025932995  0.2542892297
#> 19   Gender.Men 3  0.015671552  0.2090916856
#> 28   Gender.Men 4 -0.039478096  0.2643847300
#> 37   Gender.Men 5  0.048576225  0.2747342766
#> 2  Gender.Women 1  0.068024089 -0.2525196540
#> 11 Gender.Women 2 -0.001478645 -0.2014283024
#> 20 Gender.Women 3  0.028448294 -0.2475898965
#> 29 Gender.Women 4  0.025143980 -0.2157982545
#> 38 Gender.Women 5  0.022144648 -0.2457849931