Bootstrap validation for MCA (active variables)

Bootstrap validation of MCA, through the computation of the coordinates of active variables for bootstrap replications of the data.

bootvalid_variables(resmca, axes = c(1,2), type = "partial", K = 30)

Arguments

resmca

object of class speMCA.

axes

numeric vector of length 2, specifying the components (axes) to plot. Default is c(1,2).

type

character string. Can be "partial", "total1", "total2" or "total3" (see details). Default is "partial".

K

integer. Number of bootstrap replications (default is 30).

Details

The bootstrap technique is used here as an internal (and non-parametric) validation procedure of the results of a multiple correspondence analysis. Following the work of Lebart, several methods are proposed. The total bootstrap uses new MCAs computed from bootstrap replications of the initial data. In the type 1 bootstrap (type = "total1"), the sign of the coordinates is corrected if necessary (the direction of the axes of an ACM being arbitrary). In type 2 (type = "total2"), the order of the axes and the sign of the coordinates are corrected if necessary. In type 3 (type = "total3"), a procrustean rotation is used to find the best superposition between the initial axes and the replicated axes. The partial bootstrap (type = "partial") does not compute new MCAs: it projects bootstrap replications of the initial data as supplementary elements of the MCA. It gives a more optimistic view of the stability of the results than the total bootstrap. It is also faster. See references for more details, pros and cons of the various types, etc.

Value

A data frame with the following elements :

varcat

Name of the active category

K

Index of the bootstrap replication

dim.x

Bootstrap coordinate on the first selected axis

dim.y

Bootstrap coordinate on the second selected axis

References

Lebart L. (2006). "Validation Techniques in Multiple Correspondence Analysis". In M. Greenacre et J. Blasius (eds), Multiple Correspondence Analysis and related techniques, Chapman and Hall/CRC, p.179-196.

Lebart L. (2007). "Which bootstrap for principal axes methods?". In P. Brito et al. (eds), Selected Contributions in Data Analysis and Classification, Springer, p.581-588.

Author

Nicolas Robette

See also

Examples

data(Taste)
resmca <- speMCA(Taste[,1:11], excl=c(3,6,9,12,15,18,21,24,27,30,33))
bootvalid_variables(resmca, type = "partial", K = 5)
#>            varcat K        dim.1        dim.2
#> 15   Animation.No 1  0.042717899 -0.057906501
#> 37   Animation.No 2  0.017020103 -0.079275389
#> 59   Animation.No 3 -0.007903677 -0.065144623
#> 81   Animation.No 4  0.012042819 -0.032013655
#> 103  Animation.No 5  0.007379320 -0.063054189
#> 16  Animation.Yes 1 -0.511263272  1.023107744
#> 38  Animation.Yes 2 -0.675926727  1.057836628
#> 60  Animation.Yes 3 -0.429839298  1.017757613
#> 82  Animation.Yes 4 -0.587679182  1.157311931
#> 104 Animation.Yes 5 -0.591956173  1.164590956
#> 9    Classical.No 1  0.427618390  0.117156193
#> 31   Classical.No 2  0.443750024  0.057439404
#> 53   Classical.No 3  0.403821291  0.115445371
#> 75   Classical.No 4  0.444862673  0.105482726
#> 97   Classical.No 5  0.397743273  0.128913611
#> 10  Classical.Yes 1 -1.088273901 -0.286083706
#> 32  Classical.Yes 2 -1.108455013 -0.218417089
#> 54  Classical.Yes 3 -1.110291559 -0.300190203
#> 76  Classical.Yes 4 -1.137868533 -0.215963125
#> 98  Classical.Yes 5 -1.136350878 -0.398019114
#> 11      Comedy.No 1 -0.428327372  0.023084158
#> 33      Comedy.No 2 -0.451163800 -0.002820847
#> 55      Comedy.No 3 -0.453718834 -0.023364760
#> 77      Comedy.No 4 -0.487188714  0.011661767
#> 99      Comedy.No 5 -0.424405440 -0.036600583
#> 12     Comedy.Yes 1  0.595798262 -0.014823132
#> 34     Comedy.Yes 2  0.557024505 -0.046877546
#> 56     Comedy.Yes 3  0.567972980  0.021206972
#> 78     Comedy.Yes 4  0.592927819  0.017845730
#> 100    Comedy.Yes 5  0.567525669  0.020292942
#> 13       Crime.No 1  0.128934036 -0.017946491
#> 35       Crime.No 2  0.108140375 -0.048872438
#> 57       Crime.No 3  0.115841405 -0.021001644
#> 79       Crime.No 4  0.114386582 -0.009046679
#> 101      Crime.No 5  0.117415353 -0.029198564
#> 14      Crime.Yes 1 -0.306919170  0.074977117
#> 36      Crime.Yes 2 -0.352512732  0.052343508
#> 58      Crime.Yes 3 -0.420185645  0.037945833
#> 80      Crime.Yes 4 -0.331831102  0.073677476
#> 102     Crime.Yes 5 -0.352493446  0.033093994
#> 1    FrenchPop.No 1  0.240100902  0.589124945
#> 23   FrenchPop.No 2  0.180791591  0.562536789
#> 45   FrenchPop.No 3  0.225871988  0.591699465
#> 67   FrenchPop.No 4  0.234563266  0.574200502
#> 89   FrenchPop.No 5  0.234826082  0.583310216
#> 2   FrenchPop.Yes 1 -0.135672710 -0.357378233
#> 24  FrenchPop.Yes 2 -0.137097886 -0.372001992
#> 46  FrenchPop.Yes 3 -0.193169390 -0.369572684
#> 68  FrenchPop.Yes 4 -0.174648374 -0.328830630
#> 90  FrenchPop.Yes 5 -0.170821769 -0.371443444
#> 7         Jazz.No 1  0.317183648 -0.063599309
#> 29        Jazz.No 2  0.302504407 -0.097732906
#> 51        Jazz.No 3  0.283463978 -0.052404779
#> 73        Jazz.No 4  0.329329463 -0.065489710
#> 95        Jazz.No 5  0.306548563 -0.074163057
#> 8        Jazz.Yes 1 -1.454050037  0.325894228
#> 30       Jazz.Yes 2 -1.483151530  0.305105912
#> 52       Jazz.Yes 3 -1.446003956  0.193214261
#> 74       Jazz.Yes 4 -1.444255894  0.327927608
#> 96       Jazz.Yes 5 -1.462035926  0.242627421
#> 19        Love.No 1 -0.038146723  0.176006608
#> 41        Love.No 2 -0.066200348  0.162886236
#> 63        Love.No 3 -0.072915803  0.170031058
#> 85        Love.No 4 -0.062942143  0.198202211
#> 107       Love.No 5 -0.040634858  0.166099205
#> 20       Love.Yes 1  0.412200299 -1.364490650
#> 42       Love.Yes 2  0.380602890 -1.436713566
#> 64       Love.Yes 3  0.334132439 -1.430020543
#> 86       Love.Yes 4  0.417646551 -1.433931339
#> 108      Love.Yes 5  0.208779706 -1.505240513
#> 21     Musical.No 1  0.041514417  0.082842358
#> 43     Musical.No 2  0.025106624  0.067738986
#> 65     Musical.No 3  0.014118416  0.080447410
#> 87     Musical.No 4  0.031462402  0.098819069
#> 109    Musical.No 5  0.038349878  0.098879579
#> 22    Musical.Yes 1 -1.073633985 -2.466646974
#> 44    Musical.Yes 2 -1.145428278 -2.432216526
#> 66    Musical.Yes 3 -1.350720048 -2.347862250
#> 88    Musical.Yes 4 -1.154254165 -2.258911450
#> 110   Musical.Yes 5 -1.188130428 -2.463613812
#> 3          Rap.No 1 -0.062432318 -0.141034322
#> 25         Rap.No 2 -0.084230217 -0.171981850
#> 47         Rap.No 3 -0.111278436 -0.164944847
#> 69         Rap.No 4 -0.098807782 -0.182969212
#> 91         Rap.No 5 -0.106282637 -0.178635033
#> 4         Rap.Yes 1  0.549367408  1.053521325
#> 26        Rap.Yes 2  0.470919292  0.983031886
#> 48        Rap.Yes 3  0.537662804  1.082507671
#> 70        Rap.Yes 4  0.459870375  1.095091936
#> 92        Rap.Yes 5  0.561062071  1.043030192
#> 5         Rock.No 1  0.166212005 -0.269985535
#> 27        Rock.No 2  0.169875670 -0.336315388
#> 49        Rock.No 3  0.129362439 -0.320231421
#> 71        Rock.No 4  0.180568529 -0.300175663
#> 93        Rock.No 5  0.147231621 -0.323019152
#> 6        Rock.Yes 1 -0.425614924  0.812153379
#> 28       Rock.Yes 2 -0.534501345  0.844727726
#> 50       Rock.Yes 3 -0.469887397  0.839833802
#> 72       Rock.Yes 4 -0.505980449  0.825604513
#> 94       Rock.Yes 5 -0.448590735  0.800868766
#> 17       SciFi.No 1  0.059801783 -0.110086825
#> 39       SciFi.No 2  0.044921630 -0.139182910
#> 61       SciFi.No 3  0.013901659 -0.107047535
#> 83       SciFi.No 4  0.045186908 -0.098944801
#> 105      SciFi.No 5  0.023277345 -0.122822830
#> 18      SciFi.Yes 1 -0.591814712  1.457364018
#> 40      SciFi.Yes 2 -0.755902759  1.458116297
#> 62      SciFi.Yes 3 -0.637094037  1.367684048
#> 84      SciFi.Yes 4 -0.766078794  1.525487336
#> 106     SciFi.Yes 5 -0.557363970  1.435187883