Computes a Burt table

Computes a Burt table from a data frame composed of categorical variables.

burt(data)

Arguments

data
data frame with n rows (individuals) and p columns (categorical variables)

Details

A Burt table is a symmetric table that is used in correspondence analysis. It shows the frequencies for all combinations of categories of pairs of variables.

Value

Returns a square matrix. Its dimension is equal to the total number of categories in the data frame.

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Burt table of variables in columns 1 to 5
## in the 'Music' example data set
data(Music)
burt(Music[,1:5])

#>               FrenchPop.No FrenchPop.Yes FrenchPop.NA Rap.No Rap.Yes Rap.NA
#> FrenchPop.No           185             0            0    155      27      3
#> FrenchPop.Yes            0           305            0    267      34      4
#> FrenchPop.NA             0             0           10      9       0      1
#> Rap.No                 155           267            9    431       0      0
#> Rap.Yes                 27            34            0      0      61      0
#> Rap.NA                   3             4            1      0       0      8
#> Rock.No                136           218            5    318      34      7
#> Rock.Yes                42            84            5    107      23      1
#> Rock.NA                  7             3            0      6       4      0
#> Jazz.No                155           245            7    358      43      6
#> Jazz.Yes                27            56            3     66      18      2
#> Jazz.NA                  3             4            0      7       0      0
#> Classical.No           132           200            6    284      48      6
#> Classical.Yes           49           100            4    140      11      2
#> Classical.NA             4             5            0      7       2      0
#>               Rock.No Rock.Yes Rock.NA Jazz.No Jazz.Yes Jazz.NA Classical.No
#> FrenchPop.No      136       42       7     155       27       3          132
#> FrenchPop.Yes     218       84       3     245       56       4          200
#> FrenchPop.NA        5        5       0       7        3       0            6
#> Rap.No            318      107       6     358       66       7          284
#> Rap.Yes            34       23       4      43       18       0           48
#> Rap.NA              7        1       0       6        2       0            6
#> Rock.No           359        0       0     305       49       5          237
#> Rock.Yes            0      131       0      97       33       1           95
#> Rock.NA             0        0      10       5        4       1            6
#> Jazz.No           305       97       5     407        0       0          299
#> Jazz.Yes           49       33       4       0       86       0           34
#> Jazz.NA             5        1       1       0        0       7            5
#> Classical.No      237       95       6     299       34       5          338
#> Classical.Yes     116       34       3     100       51       2            0
#> Classical.NA        6        2       1       8        1       0            0
#>               Classical.Yes Classical.NA
#> FrenchPop.No             49            4
#> FrenchPop.Yes           100            5
#> FrenchPop.NA              4            0
#> Rap.No                  140            7
#> Rap.Yes                  11            2
#> Rap.NA                    2            0
#> Rock.No                 116            6
#> Rock.Yes                 34            2
#> Rock.NA                   3            1
#> Jazz.No                 100            8
#> Jazz.Yes                 51            1
#> Jazz.NA                   2            0
#> Classical.No              0            0
#> Classical.Yes           153            0
#> Classical.NA              0            9