Measures the association between a continuous variable and some continuous and/or categorical variables

Measures the association between a continuous variable and some continuous and/or categorical variables

condesc(y,x,weights=rep(1,length(y)),min.cor=NULL,nperm=10,distrib="asympt")

Arguments

y
the continuous variable to describe

x
a data frame with continuous and/or categorical variables

weights
an optional numeric vector of weights (by default, a vector of 1 for uniform weights)

min.cor
for the relationship between y and a categorical variable, only associations higher or equal to min.cor will be displayed. If NULL (default), they are all displayed.

nperm
numeric. Number of permutations for the permutation test of independence. If NULL, no permutation test is performed.

distrib
the null distribution of permutation test of independence can be approximated by its asymptotic distribution ("asympt", default) or via Monte Carlo resampling ("approx").

Value

A list of the following items :

variables

associations between y and the variables in x

categories

a data frame with categorical variables from x and associations measured by correlation coefficients

References

Rakotomalala R., 'Comprendre la taille d'effet (effect size)', [http://eric.univ-lyon2.fr/~ricco/cours/slides/effect_size.pdf]

Author

Nicolas Robette

See also

Examples

data(Taste)
getindexcat(Taste[,1:5])

#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA" 

mca <- speMCA(Taste[,1:5],excl=c(3,6,9,12,15))
condesc(mca$ind$coord[,1], Taste[,c('Gender','Age')])

#> [1] 1
#> [1] 2
#> [1] 3
#> [1] 4
#> [1] 5

#> $variables
#>   variable measure association permutation.pvalue
#> 1      Age    Eta2       0.034              0.000
#> 2   Gender    Eta2       0.000              0.843
#> 
#> $categories
#>     categories median.y.in.xcat median.y.global mad.y.in.xcat mad.y.global
#> 5      Age.50+       -0.1008178      -0.1871743     0.3902661    0.2705048
#> 2 Gender.Women       -0.1950406      -0.1871743     0.2605245    0.2705048
#> 1   Gender.Men       -0.1772084      -0.1871743     0.2818235    0.2705048
#> 4    Age.25-49       -0.2050335      -0.1871743     0.1675050    0.2705048
#> 3    Age.15-24       -0.3494704      -0.1871743     0.2042172    0.2705048
#>       cor.coef
#> 5  0.157576723
#> 2  0.007807668
#> 1 -0.007807668
#> 4 -0.054660413
#> 3 -0.146839406
#>