Computes contributions for a correspondence analysis

From MCA results, computes contributions of categories and variables to the axes and the overall cloud.

contrib(resmca)

Arguments

resmca
object of class 'MCA', 'speMCA' or 'csMCA'

Details

The contribution of a point to an axis depends both on the distance from the point to the origin point along the axis and on the weight of the point. The contributions of points to axes are the main aid to interpretation (see Le Roux and Rouanet, 2004 and 2010).

Value

A list of data frames:

ctr

Data frame with the contributions of categories to axes

var.ctr

Data frame with the contributions of variables to axes

ctr.cloud

Data frame with the contributions of categories to the overall cloud

vctr.cloud

Data frame with the contributions of variables to the overall cloud

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on the 'Music' example data set
## and compute contributions
data(Music)
mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
contrib(mca)

#> $ctr
#>                   dim.1     dim.2     dim.3     dim.4     dim.5
#> FrenchPop.No   1.761939 14.230710 47.672439  0.804339  0.814737
#> FrenchPop.Yes  0.865709  7.407133 25.634429  0.402064  0.392772
#> Rap.No         0.479114  3.858544  0.434361  4.574172  0.232650
#> Rap.Yes        4.312698 37.001688  5.289258 41.525658  2.294931
#> Rock.No        4.526143  3.734124  1.535924 11.330431  2.213373
#> Rock.Yes      15.859618 12.980665  5.344875 36.340092  6.148427
#> Jazz.No        7.123092  0.003974  0.768658  0.022069  8.269819
#> Jazz.Yes      37.320968  0.014553  4.525617  0.176143 41.765072
#> Classical.No   7.016865  5.109088  2.450361  1.079650  9.646189
#> Classical.Yes 20.733854 15.659522  6.344077  3.745381 28.222031
#> 
#> $var.ctr
#>               dim.1     dim.2     dim.3     dim.4     dim.5
#> FrenchPop  2.627648 21.637843 73.306868  1.206403  1.207509
#> Rap        4.791812 40.860232  5.723619 46.099830  2.527581
#> Rock      20.385761 16.714789  6.880799 47.670523  8.361800
#> Jazz      44.444060  0.018527  5.294275  0.198212 50.034891
#> Classical 27.750719 20.768610  8.794438  4.825031 37.868220
#> 
#> $ctr.cloud
#>               ctr.cloud
#> FrenchPop.No      13.08
#> FrenchPop.Yes      6.96
#> Rap.No             2.20
#> Rap.Yes           18.12
#> Rock.No            4.92
#> Rock.Yes          15.48
#> Jazz.No            3.32
#> Jazz.Yes          16.80
#> Classical.No       5.20
#> Classical.Yes     15.00
#> 
#> $vctr.cloud
#>           vctr.cloud
#> FrenchPop      20.04
#> Rap            20.32
#> Rock           20.40
#> Jazz           20.12
#> Classical      20.20
#>