Describes the contributions to axes for MCA and variants of MCA

Identifies the categories and individuals that contribute the most to each dimension obtained by a Multiple Correspondence Analysis. It allows to analyze variants of MCA, such as 'specific' MCA or 'class specific' MCA.

dimcontrib(resmca, dim = c(1,2), best = TRUE)

Arguments

resmca
object of class 'MCA', 'speMCA', or 'csMCA'

dim
dimensions to describe (default is c(1,2))

best
if FALSE, displays all the categories; if TRUE (default), displays only categories and individuals with contributions higher than average

Details

Contributions are sorted and assigned a positive or negative sign according to the corresponding categories or individuals' coordinates, so as to facilitate interpretation.

Value

Returns a list:

var

a list of categories' contributions to axes

ind

a list of individuals' contributions to axes

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then describes the contributions to axes.
data(Music)
getindexcat(Music[,1:5])

#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA" 

mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
dimcontrib(mca)

#> $var
#> $var$axe.1
#>                  ctr weight
#> Jazz.Yes      -42.28     86
#> Classical.Yes -23.76    153
#> Classical.No   10.62    338
#> 
#> $var$axe.2
#>                  ctr weight
#> Classical.Yes -14.14    153
#> Rock.Yes       21.86    131
#> Rap.Yes        39.22     61
#> 
#> 
#> $ind
#> $ind$axe.1
#>      2442      3952      3411      4537      1147      4852      2510      4147 
#> -1.794617 -1.794617 -1.794617 -1.794617 -1.794617 -1.794617 -1.382791 -1.382791 
#>      3432       137      1204      3413      1634      1920      3422      1386 
#> -1.382791 -1.382791 -1.382791 -1.267833 -1.222992 -1.221623 -1.013878 -0.966775 
#>      1522      3722       818       516      3035      2841      1805       891 
#> -0.966775 -0.966775 -0.956391 -0.886534 -0.886534 -0.886534 -0.886534 -0.886534 
#>      3754      1321      3836        79      3121      1978       248      1648 
#> -0.886534 -0.886534 -0.886534 -0.886534 -0.886534 -0.886534 -0.886534 -0.886534 
#>      2781      2145      2433      2181      4602      4538      2106      3273 
#> -0.886534 -0.886534 -0.886534 -0.886534 -0.886534 -0.886534 -0.886534 -0.759560 
#>      3177       707       569      4451      1185      4012      2535      4092 
#> -0.709582 -0.709582 -0.709582 -0.709582 -0.560828 -0.560828 -0.560828 -0.560828 
#>      1381      1558      2596      3498       385      1857      1060       500 
#> -0.560828 -0.560828 -0.560828 -0.560828 -0.560828 -0.560828 -0.560828 -0.460573 
#>       460      4316      2406      4616      4972      2113      1712      3933 
#> -0.460573 -0.460573 -0.460573 -0.460573 -0.460573 -0.460573 -0.460573 -0.453415 
#>      4903      3994      4818      1697      2469       913       424      2302 
#> -0.422101 -0.369670 -0.369670 -0.358304 -0.236178 -0.236178 -0.236178 -0.236178 
#>      4281      2769      2376      1046       432      2547      4615      2432 
#> -0.231059  0.217739  0.217739  0.236893  0.236893  0.236893  0.236893  0.236893 
#>      1868      3481       675      1841      1736      2786      3524       724 
#>  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893 
#>      1150       692      4330      3374      1573      4014      1192       534 
#>  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893 
#>       545      3024      4270      3006      2216      2616      1625      2901 
#>  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893 
#>      2007        56       679      1277      3140      2919      2654       944 
#>  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893 
#>      1793       414      2393      4006       945       230      3428      4437 
#>  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893 
#>       440      1187      4609      1065      4033      2718      4350      2315 
#>  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893 
#>      3194      2678      4501       586      3610       375      1532      4860 
#>  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893 
#>      1842      3332      3199       198      2747      1057      3923      2600 
#>  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893  0.236893 
#>       150      4196 
#>  0.236893  0.236893 
#> 
#> $ind$axe.2
#>      4589      3865      4387      2290      1050       658      2940      3392 
#> -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 
#>      3021      1412      1983      4542      1416      4200      2356       280 
#> -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 
#>      4035        66       180      1894        49      2115      1444      3957 
#> -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 
#>      3793      4425      1105      1804      3638      4376      1338       927 
#> -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 
#>      2855      3342      3063       595      2333       290      1519       202 
#> -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 
#>      1877       157      1397      3512      2032      1696      4458      1119 
#> -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 
#>      3278      3064      3464      2735      1173        96      2352       516 
#> -0.357353 -0.357353 -0.357353 -0.357353 -0.357353 -0.337429 -0.244202 -0.241941 
#>      3035      2841      1805       891      3754      1321      3836        79 
#> -0.241941 -0.241941 -0.241941 -0.241941 -0.241941 -0.241941 -0.241941 -0.241941 
#>      3121      1978       248      1648      2781      2145      2433      2181 
#> -0.241941 -0.241941 -0.241941 -0.241941 -0.241941 -0.241941 -0.241941 -0.241941 
#>      4602      4538      2106      3901      2752      4776      1439      2941 
#> -0.241941 -0.241941 -0.241941 -0.205945  0.230313  0.230313  0.230313  0.230313 
#>      1818      4992      1589      4223      1937      4821       823      3911 
#>  0.230313  0.230313  0.230313  0.230313  0.230313  0.230313  0.230313  0.230313 
#>       835       727      2379      3145      3470      4601      1266      3860 
#>  0.230313  0.230313  0.230313  0.230313  0.230313  0.230313  0.230313  0.230313 
#>      1363       972      2469       913       424      2302      2439      1399 
#>  0.230313  0.230313  0.343190  0.343190  0.343190  0.343190  0.416417  0.416417 
#>      1616      4088      3519      1378       883      1936      2832       818 
#>  0.416417  0.416417  0.416417  0.416417  0.416417  0.416417  0.416417  0.418564 
#>      2504      3994      4818      2048      3459      4666      1664       936 
#>  0.457065  0.564330  0.564330  0.640904  0.640904  0.640904  0.640904  0.640904 
#>      1788      4025      4811       456      3043      4668      4444       459 
#>  0.640904  0.640904  0.640904  0.640904  0.640904  0.640904  0.640904  0.640904 
#>      2980      1721      4785      2661      3372      2442      3952      3411 
#>  0.640904  0.640904  0.640904  0.640904  0.672474  0.751121  0.751121  0.751121 
#>      4537      1147      4852      3933      4721      4281      1310       808 
#>  0.751121  0.751121  0.751121  0.801379  0.821705  1.103463  1.355479  1.355479 
#>      2904       637      3681       699      2961      4381      2737      3177 
#>  1.355479  1.355479  1.355479  1.355479  1.355479  1.355479  1.355479  1.613320 
#>       707       569      4451      2001      3746      1979      4903 
#>  1.613320  1.613320  1.613320  1.741109  1.741109  1.741109  2.031839 
#> 
#>