Computes a homogeneity test for a categorical supplementary variable

From MCA results, computes a homogeneity test for a categorical supplementary variable, i.e. characterizes the homogeneity of several subclouds.

homog.test(resmca, var)

Arguments

resmca
object of class 'MCA', 'speMCA', 'csMCA', 'stMCA' or 'multiMCA'

var
the categorical supplementary variable. It does not need to have been used at the MCA step.

Value

Returns a list of ncp lists, with ncp the number of dimensions in the MCA results. Each list has 2 elements :

test.stat

The square matrix of test statistics

p.values

The square matrix of p.values

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then computes a homogeneity test for age supplementary variable.
data(Music)
getindexcat(Music)

#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA"  "Gender.Men"   
#> [17] "Gender.Women"  "Age.15-24"     "Age.25-49"     "Age.50+"      
#> [21] "OnlyMus.Daily" "OnlyMus.Often" "OnlyMus.Rare"  "OnlyMus.Never"
#> [25] "Daily.No"      "Daily.Yes"    

mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
homog.test(mca,Music$Age)

#> $dim.1
#> $dim.1$test.stat
#>          15-24    25-49      50+
#> 15-24 0.000000 7.058897 8.008692
#> 25-49 7.058897 0.000000 1.200371
#> 50+   8.008692 1.200371 0.000000
#> 
#> $dim.1$p.values
#>              15-24        25-49          50+
#> 15-24 1.000000e+00 1.678213e-12 1.110223e-15
#> 25-49 1.678213e-12 1.000000e+00 2.299954e-01
#> 50+   1.110223e-15 2.299954e-01 1.000000e+00
#> 
#> 
#> $dim.2
#> $dim.2$test.stat
#>           15-24     25-49      50+
#> 15-24  0.000000  8.772764 20.80055
#> 25-49  8.772764  0.000000 16.18387
#> 50+   20.800554 16.183873  0.00000
#> 
#> $dim.2$p.values
#>       15-24 25-49 50+
#> 15-24     1     0   0
#> 25-49     0     1   0
#> 50+       0     0   1
#> 
#> 
#> $dim.3
#> $dim.3$test.stat
#>          15-24    25-49      50+
#> 15-24 0.000000 6.974926 5.490793
#> 25-49 6.974926 0.000000 2.095197
#> 50+   5.490793 2.095197 0.000000
#> 
#> $dim.3$p.values
#>              15-24        25-49          50+
#> 15-24 1.000000e+00 3.060219e-12 4.001328e-08
#> 25-49 3.060219e-12 1.000000e+00 3.615346e-02
#> 50+   4.001328e-08 3.615346e-02 1.000000e+00
#> 
#> 
#> $dim.4
#> $dim.4$test.stat
#>          15-24    25-49      50+
#> 15-24 0.000000 1.905077 1.708058
#> 25-49 1.905077 0.000000 4.916974
#> 50+   1.708058 4.916974 0.000000
#> 
#> $dim.4$p.values
#>            15-24        25-49          50+
#> 15-24 1.00000000 5.677011e-02 8.762565e-02
#> 25-49 0.05677011 1.000000e+00 8.789226e-07
#> 50+   0.08762565 8.789226e-07 1.000000e+00
#> 
#> 
#> $dim.5
#> $dim.5$test.stat
#>          15-24    25-49      50+
#> 15-24 0.000000 2.115627 3.088210
#> 25-49 2.115627 0.000000 1.291504
#> 50+   3.088210 1.291504 0.000000
#> 
#> $dim.5$p.values
#>             15-24      25-49         50+
#> 15-24 1.000000000 0.03437653 0.002013663
#> 25-49 0.034376528 1.00000000 0.196529083
#> 50+   0.002013663 0.19652908 1.000000000
#> 
#>