Computes statistics for supplementary individuals

From MCA results, computes statistics (coordinates, squared cosines) for supplementary individuals.

indsup(resmca, supdata)

Arguments

resmca
object of class 'MCA', 'speMCA' or 'csMCA'

supdata
data frame with the supplementary individuals. It must have the same factors as the data frame used as input for the initial MCA.

Value

Returns a list:

coord

matrix of individuals' coordinates

cos2

matrix of individuals' square cosines

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then computes statistics for supplementary individuals.
data(Music)
getindexcat(Music)

#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA"  "Gender.Men"   
#> [17] "Gender.Women"  "Age.15-24"     "Age.25-49"     "Age.50+"      
#> [21] "OnlyMus.Daily" "OnlyMus.Often" "OnlyMus.Rare"  "OnlyMus.Never"
#> [25] "Daily.No"      "Daily.Yes"    

mca <- speMCA(Music[3:nrow(Music),1:5],excl=c(3,6,9,12,15))
indsup(mca,Music[1:2,1:5])

#> $coord
#>            dim.1      dim.2      dim.3      dim.4      dim.5
#> 2124 -0.03021772 -0.3899678  0.2212868  0.6879359 -0.1936124
#> 4485 -0.55418126 -0.6285349 -0.7238657 -0.5821206  0.1128835
#> 
#> $cos2
#>         dim.1    dim.2    dim.3    dim.4    dim.5
#> 2124 0.001281 0.213379 0.068708 0.664035 0.052597
#> 4485 0.194654 0.250390 0.332104 0.214775 0.008076
#>