Benzecri's modified rates of variance

Computes Benzecri's modified rates of variance of a multiple correspondence analysis.

modif.rate(resmca)

Arguments

resmca

object of class MCA, speMCA, csMCA, stMCA or multiMCA

Details

As MCA clouds often have a high dimensionality, the variance rates of the first principle axes may be quite low, which makes them hard to interpret. Benzecri (1992, p.412) proposed to use modified rates to better appreciate the relative importance of the principal axes.

Value

Returns a data frame with 2 variables:

mrate

modified rates

cum.mrate

cumulative modified rates

References

Benzecri J.P., Correspondence analysis handbook, New-York: Dekker (1992).

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

# MCA of Music' example data set
data(Music)
mca <- speMCA(Music[,1:5])
# modified rates of variance
modif.rate(mca)
#> $raw
#>           eigen         rate  cum.rate
#> 1  2.815811e-01 1.407906e+01  14.07906
#> 2  2.388327e-01 1.194164e+01  26.02069
#> 3  2.242049e-01 1.121024e+01  37.23094
#> 4  2.167211e-01 1.083606e+01  48.06699
#> 5  2.016064e-01 1.008032e+01  58.14731
#> 6  1.942821e-01 9.714105e+00  67.86142
#> 7  1.816117e-01 9.080585e+00  76.94200
#> 8  1.691445e-01 8.457224e+00  85.39922
#> 9  1.658990e-01 8.294950e+00  93.69417
#> 10 1.261165e-01 6.305826e+00 100.00000
#> 11 2.765345e-30 1.382672e-28 100.00000
#> 12 2.293643e-30 1.146822e-28 100.00000
#> 13 1.376840e-30 6.884199e-29 100.00000
#> 14 1.109218e-30 5.546091e-29 100.00000
#> 15 7.212848e-32 3.606424e-30 100.00000
#> 
#> $modif
#>         mrate cum.mrate
#> 1 73.69174730  73.69175
#> 2 16.69687606  90.38862
#> 3  6.48702831  96.87565
#> 4  3.09577754  99.97143
#> 5  0.02857078 100.00000
#>