Computes the phi coefficient for every cells of a contingency table

Computes the phi coefficient for every cells of the cross-tabulation between two categorical variables

phi.table(x,y,dec=3)

Arguments

x
the first categorical variable

y
the second categorical variable

dec
integer. The number of decimals (default is 3).

Value

A table with the phi coefficients

References

Rakotomalala R., 'Comprendre la taille d'effet (effect size)', http://eric.univ-lyon2.fr/~ricco/cours/slides/effect_size.pdf

Author

Nicolas Robette

See also

Examples

data(Music)
phi.table(Music$Jazz,Music$Age)

#>      15-24  25-49    50+
#> No   0.024  0.015 -0.032
#> Yes -0.018 -0.024  0.036
#> NA  -0.033  0.042 -0.018