Transforms a (possibly weighted) contingency table into percentages

Computes a contingency table from one or two vectors, with the possibility of specifying weights, and then computes the percentages.

prop.wtable(var1,var2=NULL,w=rep.int(1,length(var1)),dir=0,digits=1,mar=TRUE,na=TRUE)

Arguments

var1
an object which can be interpreted as factor

var2
an optional object which can be interpreted as factor

w
an optional numeric vector of weights (by default, a vector of 1 for uniform weights)

dir
integer: 0 for percentages, 1 for row percentages and 2 for column percentages (default is 0)

digits
integer indicating the number of decimal places (default is 1)

mar
logical. If TRUE (default), margins are computed

na
logical. If TRUE (default), 'NA' are treated as a category. If FALSE, they are ignored

Value

Returns a contingency table expressed in percentages in matrix format.

Author

Nicolas Robette

See also

Examples

## Computes a contingency table
## (expressed in percentages)
## of jazz and age variables
## from the 'Music' example data set
## with or without weights
data(Music)
prop.wtable(Music$Jazz)

#>      [,1]
#> No   83.4
#> Yes  16.0
#> NA    0.6
#> tot 100.0

prop.wtable(Music$Jazz,Music$Gender)

#>      Men Women NA   tot
#> No  37.6  45.8  0  83.4
#> Yes  7.2   8.8  0  16.0
#> NA   0.2   0.4  0   0.6
#> tot 45.0  55.0  0 100.0

prop.wtable(Music$Jazz,Music$Gender,dir=1)

#>      Men Women NA tot
#> No  45.1  54.9  0 100
#> Yes 45.0  55.0  0 100
#> NA  33.3  66.7  0 100
#> tot 45.0  55.0  0 100

prop.wtable(Music$Jazz,Music$Gender,dir=2)

#>       Men Women  NA   tot
#> No   83.6  83.3 NaN  83.4
#> Yes  16.0  16.0 NaN  16.0
#> NA    0.4   0.7 NaN   0.6
#> tot 100.0 100.0 NaN 100.0

weight <- rep(c(0,0.5,1,1.5,2), length.out=nrow(Music))
prop.wtable(Music$Jazz,w=weight)

#>      [,1]
#> No   83.5
#> Yes  15.9
#> NA    0.6
#> tot 100.0

prop.wtable(Music$Jazz,Music$Age,weight)

#>     15-24 25-49  50+ NA   tot
#> No   13.0  31.5 39.0  0  83.5
#> Yes   2.4   6.8  6.7  0  15.9
#> NA    0.0   0.2  0.4  0   0.6
#> tot  15.4  38.5 46.1  0 100.0

prop.wtable(Music$Jazz,Music$Age,weight,dir=1)

#>     15-24 25-49  50+ NA tot
#> No   15.6  37.7 46.7  0 100
#> Yes  15.1  42.8 42.1  0 100
#> NA    0.0  33.3 66.7  0 100
#> tot  15.4  38.5 46.1  0 100

prop.wtable(Music$Jazz,Music$Age,weight,dir=2)

#>     15-24 25-49   50+  NA   tot
#> No   84.4  81.8  84.6 NaN  83.5
#> Yes  15.6  17.7  14.5 NaN  15.9
#> NA    0.0   0.5   0.9 NaN   0.6
#> tot 100.0 100.0 100.0 NaN 100.0