Scaled deviations for a categorical supplementary variable

From MCA results, computes scaled deviations between categories for a categorical supplementary variable.

scaled.dev(resmca, var)

Arguments

resmca

object of class MCA, speMCA, csMCA, stMCA or multiMCA

var

the categorical supplementary variable. It does not need to have been used at the MCA step.

Value

Returns a list with one matrix for each dimension of the MCA. Each matrix is filled with scaled deviations between the categories of the supplementary variable, for a given dimension.

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

# specific MCA of Music example data set
data(Music)
junk <- c("FrenchPop.NA", "Rap.NA", "Rock.NA", "Jazz.NA", "Classical.NA")
mca <- speMCA(Music[,1:5], excl = junk)
# computes scaled deviations for Age supplementary variable
scaled.dev(mca,Music$Age)
#> $dim.1
#>       15-24 25-49   50+
#> 15-24 0.000 0.496 0.558
#> 25-49 0.496 0.000 0.062
#> 50+   0.558 0.062 0.000
#> 
#> $dim.2
#>       15-24 25-49   50+
#> 15-24 0.000 0.546 1.283
#> 25-49 0.546 0.000 0.737
#> 50+   1.283 0.737 0.000
#> 
#> $dim.3
#>       15-24 25-49   50+
#> 15-24 0.000 0.419 0.327
#> 25-49 0.419 0.000 0.092
#> 50+   0.327 0.092 0.000
#> 
#> $dim.4
#>       15-24 25-49   50+
#> 15-24 0.000 0.105 0.093
#> 25-49 0.105 0.000 0.199
#> 50+   0.093 0.199 0.000
#> 
#> $dim.5
#>       15-24 25-49   50+
#> 15-24 0.000 0.101 0.146
#> 25-49 0.101 0.000 0.045
#> 50+   0.146 0.045 0.000
#>