Computes statistics for categorical supplementary variables

From MCA results, computes statistics (weights, coordinates, contributions, test-values, variances) for categorical supplementary variables.

varsups(resmca, vars)

Arguments

resmca

object of class MCA, speMCA, csMCA, stMCA or multiMCA

vars

A data frame of categorical supplementary variables. All these variables should be factors.

Value

Returns a list:

weight

numeric vector of categories weights

coord

data frame of categories coordinates

cos2

data frame of categories square cosine

var

a list of data frames of categories within variances, variance between and within categories and variable square correlation ratio (eta2)

typic

data frame of categories typicality test statistics

pval

data frame of categories p-values from typicality test statistics

cor

data frame of categories correlation coefficients

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then computes statistics for supplementary variables.
data(Music)
getindexcat(Music)
#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA"  "Gender.Men"   
#> [17] "Gender.Women"  "Age.15-24"     "Age.25-49"     "Age.50+"      
#> [21] "OnlyMus.Daily" "OnlyMus.Often" "OnlyMus.Rare"  "OnlyMus.Never"
#> [25] "Daily.No"      "Daily.Yes"    
mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
varsups(mca,Music[,6:9])
#> $weight
#>    Gender.Men  Gender.Women     Age.15-24     Age.25-49       Age.50+ 
#>           253           247            78           204           218 
#> OnlyMus.Daily OnlyMus.Often  OnlyMus.Rare OnlyMus.Never      Daily.No 
#>            58           121            64           257           303 
#>     Daily.Yes 
#>           197 
#> 
#> $coord
#>                   dim.1     dim.2     dim.3     dim.4     dim.5
#> Gender.Men     0.002461 -0.095251  0.044400  0.011305  0.014944
#> Gender.Women  -0.002521  0.097565 -0.045478 -0.011579 -0.015307
#> Age.15-24      0.445924 -0.781930  0.313437  0.002163  0.105164
#> Age.25-49     -0.050511 -0.236099 -0.105504 -0.102956  0.003924
#> Age.50+       -0.112283  0.500710 -0.013419  0.095570 -0.041299
#> OnlyMus.Daily -0.288376 -0.388022  0.318565  0.155123 -0.062893
#> OnlyMus.Often -0.071372 -0.176565  0.082970  0.202373  0.141510
#> OnlyMus.Rare  -0.261283 -0.138284 -0.191566 -0.206694 -0.035928
#> OnlyMus.Never  0.163751  0.205135 -0.063253 -0.078817 -0.043484
#> Daily.No       0.018874  0.246987 -0.059847  0.016209  0.000612
#> Daily.Yes     -0.029029 -0.379884  0.092050 -0.024931 -0.000941
#> 
#> $cos2
#>                  dim.1    dim.2    dim.3    dim.4    dim.5
#> Gender.Men    0.000006 0.009293 0.002019 0.000131 0.000229
#> Gender.Women  0.000006 0.009293 0.002019 0.000131 0.000229
#> Age.15-24     0.036754 0.113010 0.018159 0.000001 0.002044
#> Age.25-49     0.001758 0.038417 0.007671 0.007305 0.000011
#> Age.50+       0.009746 0.193811 0.000139 0.007061 0.001319
#> OnlyMus.Daily 0.010912 0.019757 0.013317 0.003158 0.000519
#> OnlyMus.Often 0.001626 0.009953 0.002198 0.013075 0.006393
#> OnlyMus.Rare  0.010021 0.002807 0.005387 0.006271 0.000189
#> OnlyMus.Never 0.028359 0.044505 0.004231 0.006570 0.002000
#> Daily.No      0.000548 0.093827 0.005509 0.000404 0.000001
#> Daily.Yes     0.000548 0.093827 0.005509 0.000404 0.000001
#> 
#> $var
#> $var$Gender
#>            dim.1    dim.2    dim.3    dim.4    dim.5
#> Men     0.290292 0.244867 0.207113 0.178409 0.129570
#> Women   0.266459 0.186373 0.198275 0.164282 0.128262
#> within  0.278519 0.215971 0.202747 0.171430 0.128924
#> between 0.000002 0.002026 0.000410 0.000022 0.000029
#> total   0.278521 0.217997 0.203157 0.171452 0.128953
#> eta2    0.000006 0.009293 0.002019 0.000131 0.000229
#> 
#> $var$Age
#>            dim.1    dim.2    dim.3    dim.4    dim.5
#> 15-24   0.224468 0.162922 0.216555 0.333632 0.080558
#> 25-49   0.299116 0.241927 0.199184 0.195215 0.123800
#> 50+     0.254595 0.101594 0.192788 0.087921 0.150360
#> within  0.268060 0.168417 0.199105 0.170028 0.128634
#> between 0.010461 0.049580 0.004052 0.001424 0.000319
#> total   0.278521 0.217997 0.203157 0.171452 0.128953
#> eta2    0.037558 0.227433 0.019946 0.008308 0.002475
#> 
#> $var$OnlyMus
#>            dim.1    dim.2    dim.3    dim.4    dim.5
#> Daily   0.375575 0.293178 0.189967 0.265148 0.142691
#> Often   0.320990 0.238196 0.208026 0.213203 0.159795
#> Rare    0.303565 0.224243 0.196405 0.169642 0.142544
#> Never   0.212287 0.169145 0.197543 0.123975 0.106332
#> within  0.269218 0.207296 0.199055 0.167790 0.128123
#> between 0.009303 0.010701 0.004102 0.003663 0.000831
#> total   0.278521 0.217997 0.203157 0.171452 0.128953
#> eta2    0.033400 0.049087 0.020192 0.021364 0.006442
#> 
#> $var$Daily
#>            dim.1    dim.2    dim.3    dim.4    dim.5
#> No      0.261124 0.161658 0.188060 0.136252 0.125526
#> Yes     0.304891 0.252736 0.223537 0.225417 0.134224
#> within  0.278368 0.197543 0.202038 0.171383 0.128953
#> between 0.000153 0.020454 0.001119 0.000069 0.000000
#> total   0.278521 0.217997 0.203157 0.171452 0.128953
#> eta2    0.000548 0.093827 0.005509 0.000404 0.000001
#> 
#> 
#> $typic
#>                   dim.1     dim.2     dim.3     dim.4     dim.5
#> Gender.Men     0.055648 -2.153442  1.003786  0.255574  0.337862
#> Gender.Women  -0.055648  2.153442 -1.003786 -0.255574 -0.337862
#> Age.15-24      4.282548 -7.509471  3.010171  0.020778  1.009972
#> Age.25-49     -0.936718 -4.378377 -1.956531 -1.909278  0.072764
#> Age.50+       -2.205306  9.834220 -0.263555  1.877042 -0.811141
#> OnlyMus.Daily -2.333523 -3.139856  2.577812  1.255247 -0.508929
#> OnlyMus.Often -0.900852 -2.228578  1.047240  2.554326  1.786115
#> OnlyMus.Rare  -2.236191 -1.183501 -1.639519 -1.768988 -0.307491
#> OnlyMus.Never  3.761817  4.712533 -1.453089 -1.810637 -0.998955
#> Daily.No       0.522876  6.842473 -1.657996  0.449054  0.016952
#> Daily.Yes     -0.522876 -6.842473  1.657996 -0.449054 -0.016952
#> 
#> $pval
#>                  dim.1    dim.2    dim.3    dim.4    dim.5
#> Gender.Men    0.955622 0.031284 0.315482 0.798279 0.735467
#> Gender.Women  0.955622 0.031284 0.315482 0.798279 0.735467
#> Age.15-24     0.000018 0.000000 0.002611 0.983423 0.312509
#> Age.25-49     0.348903 0.000012 0.050403 0.056226 0.941994
#> Age.50+       0.027433 0.000000 0.792123 0.060512 0.417285
#> OnlyMus.Daily 0.019621 0.001690 0.009943 0.209389 0.610802
#> OnlyMus.Often 0.367667 0.025842 0.294989 0.010639 0.074081
#> OnlyMus.Rare  0.025339 0.236611 0.101105 0.076896 0.758470
#> OnlyMus.Never 0.000169 0.000002 0.146199 0.070197 0.317817
#> Daily.No      0.601061 0.000000 0.097318 0.653393 0.986475
#> Daily.Yes     0.601061 0.000000 0.097318 0.653393 0.986475
#> 
#> $cor
#>                dim.1  dim.2  dim.3  dim.4  dim.5
#> Gender.Men     0.002 -0.096  0.045  0.011  0.015
#> Gender.Women  -0.002  0.096 -0.045 -0.011 -0.015
#> Age.15-24      0.192 -0.336  0.135  0.001  0.045
#> Age.25-49     -0.042 -0.196 -0.088 -0.085  0.003
#> Age.50+       -0.099  0.440 -0.012  0.084 -0.036
#> OnlyMus.Daily -0.104 -0.141  0.115  0.056 -0.023
#> OnlyMus.Often -0.040 -0.100  0.047  0.114  0.080
#> OnlyMus.Rare  -0.100 -0.053 -0.073 -0.079 -0.014
#> OnlyMus.Never  0.168  0.211 -0.065 -0.081 -0.045
#> Daily.No       0.023  0.306 -0.074  0.020  0.001
#> Daily.Yes     -0.023 -0.306  0.074 -0.020 -0.001
#>