Computes a (possibly weighted) contingency table

Computes a contingency table from one or two vectors, with the possibility of specifying weights.

wtable(x,y=NULL,weights=rep.int(1,length(x)),stat="freq",
              digits=1,mar=TRUE,na_value=NULL)

Arguments

x

an object which can be interpreted as factor

y

an optional object which can be interpreted as factor

weights

an optional numeric vector of weights (by default, a vector of 1 for uniform weights)

stat

character. Whether to compute a contingenct table ("freq", default, percentages ("prop"), row percentages ("rprop") or column percentages ("cprop").

digits

integer indicating the number of decimal places (default is 1)

mar

logical. If TRUE (default), margins are computed

na_value

character. Name of the level for NA category. If NULL (default), NA values are ignored.

Value

Returns a contingency table.

Author

Nicolas Robette

See also

Examples

## Computes a contingency table
## of jazz and age variables
## from the 'Music' example data set
## with or without weights
data(Music)
wtable(Music$Jazz)
#>  No Yes  NA Sum 
#> 395  95  10 500 
wtable(Music$Jazz,Music$Age)
#>     15-24 25-49 50+ Sum
#> No     67   159 169 395
#> Yes     8    43  44  95
#> NA      3     2   5  10
#> Sum    78   204 218 500

weight <- rep(c(0,0.5,1,1.5,2), length.out=nrow(Music))
wtable(Music$Jazz,w=weight,digits=1)
#>    No   Yes    NA   Sum 
#> 396.0  92.5  11.5 500.0 
wtable(Music$Jazz,Music$Age,weight,1)
#>     15-24 25-49   50+   Sum
#> No   71.0 153.0 172.0 396.0
#> Yes   7.0  41.0  44.5  92.5
#> NA    4.0   3.0   4.5  11.5
#> Sum  82.0 197.0 221.0 500.0