Cross-tabulation and measures of association between two categorical variables

Cross-tabulation and measures of association between two categorical variables

assoc.twocat(x,y,weights=rep.int(1,length(x)),na_value=NULL,nperm=1000,distrib="asympt")

Arguments

x
the first categorical variable (must be a factor)

y
the second categorical variable (must be a factor)

weights
an optional numeric vector of weights (by default, a vector of 1 for uniform weights)

na_value
character. Name of the level for NA category. If NULL (default), NA values are ignored.

nperm
numeric. Number of permutations for the permutation test of independence. If NULL, no permutation test is performed. Default is 1000.

distrib
the null distribution of permutation test of independence can be approximated by its asymptotic distribution ("asympt", default) or via Monte Carlo resampling ("approx".

Value

A list with the following elements :

freq

cross-tabulation

prop

percentages

rprop

row percentages

cprop

column percentages

expected

expected values

chi.squared

chi-squared value

cramer.v

Cramer's V between the two variables

permutation.pvalue

p-value from a permutation (so non-parametric) test of independence

pearson.residuals

the table of Pearson residuals, i.e. (observed - expected) / sqrt(expected).

phi

the table of the phi coefficients for each pair of levels

phi.perm.pval

the table of permutation p-values for each pair of levels

local.pem

the table of local Percentages of Maximum Deviation from Independences

global.pem

value of the global Percentage of Maximum Deviation from Independence

gather

a data frame gathering informations, with one row per cell of the cross-tabulation

References

Rakotomalala R., 'Comprendre la taille d'effet (effect size)', http://eric.univ-lyon2.fr/~ricco/cours/slides/effect_size.pdf

Author

Nicolas Robette

See also

Examples

data(Music)
assoc.twocat(Music$Jazz,Music$Age,nperm=100)

#> $freq
#>     15-24 25-49 50+ Sum
#> No     65   169 183 417
#> Yes    11    30  39  80
#> NA      0     2   1   3
#> Sum    76   201 223 500
#> 
#> $prop
#>     15-24 25-49   50+   Sum
#> No   13.0  33.8  36.6  83.4
#> Yes   2.2   6.0   7.8  16.0
#> NA    0.0   0.4   0.2   0.6
#> Sum  15.2  40.2  44.6 100.0
#> 
#> $rprop
#>     15-24 25-49  50+ Sum
#> No   15.6  40.5 43.9 100
#> Yes  13.8  37.5 48.8 100
#> NA    0.0  66.7 33.3 100
#> Sum  15.2  40.2 44.6 100
#> 
#> $cprop
#>     15-24 25-49   50+   Sum
#> No   85.5  84.1  82.1  83.4
#> Yes  14.5  14.9  17.5  16.0
#> NA    0.0   1.0   0.4   0.6
#> Sum 100.0 100.0 100.0 100.0
#> 
#> $expected
#>       15-24   25-49     50+
#> No   63.384 167.634 185.982
#> Yes  12.160  32.160  35.680
#> NA    0.456   1.206   1.338
#> 
#> $chi.squared
#> [1] 1.728935
#> 
#> $cramer.v
#> [1] 0.04158046
#> 
#> $permutation.pvalue
#> [1] 0.8189024
#> 
#> $pearson.residuals
#>          15-24      25-49        50+
#> No   0.2029792  0.1055042 -0.2186614
#> Yes -0.3326528 -0.3808866  0.5558091
#> NA  -0.6752777  0.7230142 -0.2922057
#> 
#> $phi
#>           15-24       25-49         50+
#> No   0.02419438  0.01497545 -0.03224618
#> Yes -0.01762660 -0.02403369  0.03643730
#> NA  -0.03289321  0.04193900 -0.01760985
#> 
#> $phi.perm.pval
#>         15-24     25-49       50+
#> No  0.3185468 0.3674241 0.2339112
#> Yes 0.3710255 0.2914890 0.2035863
#> NA  0.2547834 0.1419382 0.3015651
#> 
#> $local.pem
#>      15-24  25-49    50+
#> No    12.8    4.1   -6.5
#> Yes   -9.5   -6.7    7.5
#> NA  -100.0   44.3  -25.3
#> 
#> $global.pem
#> [1] 5.5
#> 
#> $gather
#>   Var1  Var2 Freq  prop     rprop       cprop expected std.residuals
#> 1   No 15-24   65 0.130 0.1558753 0.855263158   63.384     0.2029792
#> 2  Yes 15-24   11 0.022 0.1375000 0.144736842   12.160    -0.3326528
#> 3   NA 15-24    0 0.000 0.0000000 0.000000000    0.456    -0.6752777
#> 4   No 25-49  169 0.338 0.4052758 0.840796020  167.634     0.1055042
#> 5  Yes 25-49   30 0.060 0.3750000 0.149253731   32.160    -0.3808866
#> 6   NA 25-49    2 0.004 0.6666667 0.009950249    1.206     0.7230142
#> 7   No   50+  183 0.366 0.4388489 0.820627803  185.982    -0.2186614
#> 8  Yes   50+   39 0.078 0.4875000 0.174887892   35.680     0.5558091
#> 9   NA   50+    1 0.002 0.3333333 0.004484305    1.338    -0.2922057
#>           phi perm.pval local.pem
#> 1  0.02419438 0.3185468      12.8
#> 2 -0.01762660 0.3710255      -9.5
#> 3 -0.03289321 0.2547834    -100.0
#> 4  0.01497545 0.3674241       4.1
#> 5 -0.02403369 0.2914890      -6.7
#> 6  0.04193900 0.1419382      44.3
#> 7 -0.03224618 0.2339112      -6.5
#> 8  0.03643730 0.2035863       7.5
#> 9 -0.01760985 0.3015651     -25.3
#>