Cross-tabulation and measures of association between two categorical variables

Cross-tabulation and measures of association between two categorical variables

assoc.twocat(x,y,weights=rep.int(1,length(x)),na=TRUE,nperm=1000,distrib="asympt")

Arguments

x
the first categorical variable (must be a factor)

y
the second categorical variable (must be a factor)

weights
an optional numeric vector of weights (by default, a vector of 1 for uniform weights)

na
logical. If TRUE (default), 'NA' are treated as a category. If FALSE, they are ignored

nperm
numeric. Number of permutations for the permutation test of independence. If NULL, no permutation test is performed. Default is 1000.

distrib
the null distribution of permutation test of independence can be approximated by its asymptotic distribution ("asympt", default) or via Monte Carlo resampling ("approx".

Value

A list with the following elements :

freq

cross-tabulation

prop

percentages

rprop

raw percentages

cprop

column percentages

expected

expected values

chi.squared

chi-squared value

cramer.v

Cramer's V between the two variables

permutation.pvalue

p-value from a permutation (so non-parametric) test of independence

pearson.residuals

the table of Pearson residuals, i.e. (observed - expected) / sqrt(expected).

phi

the table of the phi coefficients for each pair of levels

phi.perm.pval

the table of permutation p-values for each pair of levels

local.pem

local percentages of maximum deviation from independence

global.pem

the global percentage of maximum deviation from independence

References

Rakotomalala R., 'Comprendre la taille d'effet (effect size)', http://eric.univ-lyon2.fr/~ricco/cours/slides/effect_size.pdf

Author

Nicolas Robette

See also

Examples

data(Music)
assoc.twocat(Music$Jazz,Music$Age)

#> $freq
#>     15-24 25-49 50+ Sum
#> No     65   169 183 417
#> Yes    11    30  39  80
#> NA      0     2   1   3
#> Sum    76   201 223 500
#> 
#> $prop
#>     15-24 25-49   50+   Sum
#> No   13.0  33.8  36.6  83.4
#> Yes   2.2   6.0   7.8  16.0
#> NA    0.0   0.4   0.2   0.6
#> Sum  15.2  40.2  44.6 100.0
#> 
#> $rprop
#>     15-24 25-49  50+ Sum
#> No   15.6  40.5 43.9 100
#> Yes  13.8  37.5 48.8 100
#> NA    0.0  66.7 33.3 100
#> Sum  15.2  40.2 44.6 100
#> 
#> $cprop
#>     15-24 25-49   50+   Sum
#> No   85.5  84.1  82.1  83.4
#> Yes  14.5  14.9  17.5  16.0
#> NA    0.0   1.0   0.4   0.6
#> Sum 100.0 100.0 100.0 100.0
#> 
#> $expected
#>          data.15-24 data.25-49 data.50+
#> data.No      63.384    167.634  185.982
#> data.Yes     12.160     32.160   35.680
#> data.NA       0.456      1.206    1.338
#> 
#> $chi.squared
#> [1] 1.728935
#> 
#> $cramer.v
#> [1] 0.04158046
#> 
#> $permutation.pvalue
#> [1] 0.8015164
#> 
#> $pearson.residuals
#>          data.15-24 data.25-49   data.50+
#> data.No   0.2029792  0.1055042 -0.2186614
#> data.Yes -0.3326528 -0.3808866  0.5558091
#> data.NA  -0.6752777  0.7230142 -0.2922057
#> 
#> $phi
#>           15-24       25-49         50+
#> No   0.02419438  0.01497545 -0.03224618
#> Yes -0.01762660 -0.02403369  0.03643730
#> NA  -0.03289321  0.04193900 -0.01760985
#> 
#> $phi.perm.pval
#>         15-24     25-49       50+
#> No  0.2717609 0.3686974 0.2215349
#> Yes 0.3208135 0.2948716 0.1945620
#> NA  0.2277871 0.1765021 0.3523111
#> 
#> $local.pem
#>          data.15-24 data.25-49 data.50+
#> data.No        12.8        4.1     -6.5
#> data.Yes       -9.5       -6.7      7.5
#> data.NA      -100.0       44.3    -25.3
#> 
#> $global.pem
#> [1] 5.5
#>