Cross-tabulation and measures of association between two categorical variables

Cross-tabulation and measures of association between two categorical variables

assoc.twocat(x,y,weights=rep.int(1,length(x)),na_value=NULL,nperm=NULL,distrib="asympt")

Arguments

x

the first categorical variable (must be a factor)

y

the second categorical variable (must be a factor)

weights

an optional numeric vector of weights (by default, a vector of 1 for uniform weights)

na_value

character. Name of the level for NA category. If NULL (default), NA values are ignored.

nperm

numeric. Number of permutations for the permutation test of independence. If NULL (default), no permutation test is performed.

distrib

the null distribution of permutation test of independence can be approximated by its asymptotic distribution (asympt, default) or via Monte Carlo resampling (approx).

Value

A list of lists with the following elements :

tables list :

freq

cross-tabulation frequencies

prop

percentages

rprop

row percentages

cprop

column percentages

expected

expected values

global list :

chi.squared

chi-squared value

cramer.v

Cramer's V between the two variables

permutation.pvalue

p-value from a permutation (so non-parametric) test of independence

global.pem

global PEM

GK.tau.xy

Goodman and Kruskal tau (forward association)

GK.tau.yx

Goodman and Kruskal tau (backward association)

local list :

std.residuals

the table of standardized (i.e.Pearson) residuals.

adj.residuals

the table of adjusted standardized residuals.

odds.ratios

the table of odds ratios.

local.pem

the table of local PEM

phi

the table of the phi coefficients for each pair of levels

phi.perm.pval

the table of permutation p-values for each pair of levels

gather : a data frame gathering informations, with one row per cell of the cross-tabulation.

References

Agresti, A. (2007). An Introduction to Categorical Data Analysis, 2nd ed. New York: John Wiley & Sons.

Rakotomalala R., Comprendre la taille d'effet (effect size), http://eric.univ-lyon2.fr/~ricco/cours/slides/effect_size.pdf

Author

Nicolas Robette

See also

Examples

data(Music)
assoc.twocat(Music$Jazz,Music$Age,nperm=100)
#> $tables
#> $tables$freq
#>     15-24 25-49 50+ Sum
#> No     67   159 169 395
#> Yes     8    43  44  95
#> NA      3     2   5  10
#> Sum    78   204 218 500
#> 
#> $tables$prop
#>     15-24 25-49   50+   Sum
#> No   13.4  31.8  33.8  79.0
#> Yes   1.6   8.6   8.8  19.0
#> NA    0.6   0.4   1.0   2.0
#> Sum  15.6  40.8  43.6 100.0
#> 
#> $tables$rprop
#>         15-24    25-49      50+ Sum
#> No  16.962025 40.25316 42.78481 100
#> Yes  8.421053 45.26316 46.31579 100
#> NA  30.000000 20.00000 50.00000 100
#> Sum 15.600000 40.80000 43.60000 100
#> 
#> $tables$cprop
#>          15-24       25-49        50+ Sum
#> No   85.897436  77.9411765  77.522936  79
#> Yes  10.256410  21.0784314  20.183486  19
#> NA    3.846154   0.9803922   2.293578   2
#> Sum 100.000000 100.0000000 100.000000 100
#> 
#> $tables$expected
#>      15-24  25-49    50+
#> No   61.62 161.16 172.22
#> Yes  14.82  38.76  41.42
#> NA    1.56   4.08   4.36
#> 
#> 
#> $global
#> $global$chi.squared
#> [1] 6.805458
#> 
#> $global$cramer.v
#> [1] 0.0824952
#> 
#> $global$permutation.pvalue
#> [1] 0.1760267
#> 
#> $global$global.pem
#> [1] 12.22184
#> 
#> $global$GK.tau.xy
#> [1] 0.004978831
#> 
#> $global$GK.tau.yx
#> [1] 0.007107588
#> 
#> 
#> $local
#> $local$std.residuals
#>          15-24      25-49        50+
#> No   0.6853642 -0.1701473 -0.2453658
#> Yes -1.7715780  0.6810421  0.4008802
#> NA   1.1529227 -1.0297534  0.3065044
#> 
#> $local$adj.residuals
#>          15-24      25-49        50+
#> No   1.6279481 -0.4825637 -0.7129595
#> Yes -2.1426268  0.9834915  0.5931064
#> NA   1.2676984 -1.3519461  0.4122723
#> 
#> $local$odds.ratios
#>         15-24     25-49       50+
#> No  1.7455654 0.8983051 0.8546144
#> Yes 0.4400657 1.2532250 1.1453685
#> NA  2.3714286 0.3564356 1.3004695
#> 
#> $local$local.pem
#>          15-24      25-49        50+
#> No   32.844933  -3.474903  -5.437352
#> Yes -46.018893   7.539118   4.815230
#> NA   17.061611 -50.980392  11.347518
#> 
#> $local$phi
#>           15-24       25-49         50+
#> No   0.07280405 -0.02158090 -0.03188452
#> Yes -0.09582119  0.04398308  0.02652452
#> NA   0.05669319 -0.06046087  0.01843738
#> 
#> $local$phi.perm.pval
#>          15-24      25-49        50+
#> No  0.06952932 0.30372654 0.25329032
#> Yes 0.02290727 0.15026083 0.28251352
#> NA  0.11497453 0.09425158 0.35691726
#> 
#> 
#> $gather
#>   var.y var.x freq  prop      rprop       cprop expected std.residuals
#> 1 15-24   Yes    8 0.016 0.08421053 0.102564103    14.82    -1.7715780
#> 2 15-24    NA    3 0.006 0.30000000 0.038461538     1.56     1.1529227
#> 3 15-24    No   67 0.134 0.16962025 0.858974359    61.62     0.6853642
#> 4 25-49    NA    2 0.004 0.20000000 0.009803922     4.08    -1.0297534
#> 5 25-49    No  159 0.318 0.40253165 0.779411765   161.16    -0.1701473
#> 6 25-49   Yes   43 0.086 0.45263158 0.210784314    38.76     0.6810421
#> 7   50+    No  169 0.338 0.42784810 0.775229358   172.22    -0.2453658
#> 8   50+    NA    5 0.010 0.50000000 0.022935780     4.36     0.3065044
#> 9   50+   Yes   44 0.088 0.46315789 0.201834862    41.42     0.4008802
#>   adj.residuals        or        pem         phi  perm.pval freq.x freq.y
#> 1    -2.1426268 0.4400657 -46.018893 -0.09582119 0.02290727     95     78
#> 2     1.2676984 2.3714286  17.061611  0.05669319 0.11497453     10     78
#> 3     1.6279481 1.7455654  32.844933  0.07280405 0.06952932    395     78
#> 4    -1.3519461 0.3564356 -50.980392 -0.06046087 0.09425158     10    204
#> 5    -0.4825637 0.8983051  -3.474903 -0.02158090 0.30372654    395    204
#> 6     0.9834915 1.2532250   7.539118  0.04398308 0.15026083     95    204
#> 7    -0.7129595 0.8546144  -5.437352 -0.03188452 0.25329032    395    218
#> 8     0.4122723 1.3004695  11.347518  0.01843738 0.35691726     10    218
#> 9     0.5931064 1.1453685   4.815230  0.02652452 0.28251352     95    218
#>   prop.x prop.y
#> 1   0.19  0.156
#> 2   0.02  0.156
#> 3   0.79  0.156
#> 4   0.02  0.408
#> 5   0.79  0.408
#> 6   0.19  0.408
#> 7   0.79  0.436
#> 8   0.02  0.436
#> 9   0.19  0.436
#>