Computes a Burt table

Computes a Burt table from a data frame composed of categorical variables.

burt(data)

Arguments

data
data frame with n rows (individuals) and p columns (categorical variables)

Details

A Burt table is a symmetric table that is used in correspondence analysis. It shows the frequencies for all combinations of categories of pairs of variables.

Value

Returns a square matrix. Its dimension is equal to the total number of categories in the data frame.

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Burt table of variables in columns 1 to 5
## in the 'Music' example data set
data(Music)
burt(Music[,1:5])

#>               FrenchPop.No FrenchPop.Yes FrenchPop.NA Rap.No Rap.Yes Rap.NA
#> FrenchPop.No           173             0            0    149      20      4
#> FrenchPop.Yes            0           326            0    295      27      4
#> FrenchPop.NA             0             0            1      1       0      0
#> Rap.No                 149           295            1    445       0      0
#> Rap.Yes                 20            27            0      0      47      0
#> Rap.NA                   4             4            0      0       0      8
#> Rock.No                131           245            1    343      27      7
#> Rock.Yes                38            75            0     92      20      1
#> Rock.NA                  4             6            0     10       0      0
#> Jazz.No                148           269            0    377      34      6
#> Jazz.Yes                25            54            1     65      13      2
#> Jazz.NA                  0             3            0      3       0      0
#> Classical.No           138           232            0    326      39      5
#> Classical.Yes           31            93            1    114       8      3
#> Classical.NA             4             1            0      5       0      0
#>               Rock.No Rock.Yes Rock.NA Jazz.No Jazz.Yes Jazz.NA Classical.No
#> FrenchPop.No      131       38       4     148       25       0          138
#> FrenchPop.Yes     245       75       6     269       54       3          232
#> FrenchPop.NA        1        0       0       0        1       0            0
#> Rap.No            343       92      10     377       65       3          326
#> Rap.Yes            27       20       0      34       13       0           39
#> Rap.NA              7        1       0       6        2       0            5
#> Rock.No           377        0       0     330       45       2          283
#> Rock.Yes            0      113       0      77       35       1           80
#> Rock.NA             0        0      10      10        0       0            7
#> Jazz.No           330       77      10     417        0       0          333
#> Jazz.Yes           45       35       0       0       80       0           34
#> Jazz.NA             2        1       0       0        0       3            3
#> Classical.No      283       80       7     333       34       3          370
#> Classical.Yes      91       31       3      80       45       0            0
#> Classical.NA        3        2       0       4        1       0            0
#>               Classical.Yes Classical.NA
#> FrenchPop.No             31            4
#> FrenchPop.Yes            93            1
#> FrenchPop.NA              1            0
#> Rap.No                  114            5
#> Rap.Yes                   8            0
#> Rap.NA                    3            0
#> Rock.No                  91            3
#> Rock.Yes                 31            2
#> Rock.NA                   3            0
#> Jazz.No                  80            4
#> Jazz.Yes                 45            1
#> Jazz.NA                   0            0
#> Classical.No              0            0
#> Classical.Yes           125            0
#> Classical.NA              0            5