Measures the association between a categorical variable and some continuous and/or categorical variables

Measures the association between a categorical variable and some continuous and/or categorical variables

catdesc(y,x,weights=rep(1,length(y)),min.phi=NULL,nperm=NULL,distrib="asympt")

Arguments

y
the categorical variable to describe (must be a factor)

x
a data frame with continuous and/or categorical variables

weights
an optional numeric vector of weights (by default, a vector of 1 for uniform weights)

min.phi
for the relationship between y and a categorical variable, only associations higher or equal to min.phi will be displayed. If NULL (default), they are all displayed.

nperm
numeric. Number of permutations for the permutation test of independence. If NULL (default), no permutation test is performed.

distrib
the null distribution of permutation test of independence can be approximated by its asymptotic distribution ("asympt", default) or via Monte Carlo resampling ("approx").

Value

A list of the following items :

variables

associations between y and the variables in x

bylevel

a list with one element for each level of y

Each element in bylevel has the following items :
categories

a data frame with categorical variables from x and associations measured by phi

continuous.var

a data frame with continuous variables from x and associations measured by correlation coefficients

References

Rakotomalala R., 'Comprendre la taille d'effet (effect size)', [http://eric.univ-lyon2.fr/~ricco/cours/slides/effect_size.pdf]

Author

Nicolas Robette

See also

Examples

data(Taste)
catdesc(Taste$Comedy, Taste[,c('Gender','Age')])

#> $variables
#>   variable    measure association permutation.pvalue
#> 1      Age Cramer's V       0.076                 NA
#> 2   Gender Cramer's V       0.015                 NA
#> 
#> $bylevel
#> $bylevel$No
#> $bylevel$No$categories
#> [1] categories       pct.ycat.in.xcat pct.xcat.in.ycat pct.xcat.global 
#> [5] phi             
#> <0 lignes> (ou 'row.names' de longueur nulle)
#> 
#> 
#> $bylevel$Yes
#> $bylevel$Yes$categories
#> [1] categories       pct.ycat.in.xcat pct.xcat.in.ycat pct.xcat.global 
#> [5] phi             
#> <0 lignes> (ou 'row.names' de longueur nulle)
#> 
#> 
#> $bylevel$`NA`
#> $bylevel$`NA`$categories
#> [1] categories       pct.ycat.in.xcat pct.xcat.in.ycat pct.xcat.global 
#> [5] phi             
#> <0 lignes> (ou 'row.names' de longueur nulle)
#> 
#> 
#>