Computes contributions for a correspondence analysis

From MCA results, computes contributions of categories and variables to the axes and the overall cloud.

contrib(resmca)

Arguments

resmca
object of class 'MCA', 'speMCA' or 'csMCA'

Details

The contribution of a point to an axis depends both on the distance from the point to the origin point along the axis and on the weight of the point. The contributions of points to axes are the main aid to interpretation (see Le Roux and Rouanet, 2004 and 2010).

Value

A list of data frames:

ctr

Data frame with the contributions of categories to axes

var.ctr

Data frame with the contributions of variables to axes

ctr.cloud

Data frame with the contributions of categories to the overall cloud

vctr.cloud

Data frame with the contributions of variables to the overall cloud

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on the 'Music' example data set
## and compute contributions
data(Music)
mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
contrib(mca)

#> $ctr
#>                   dim.1     dim.2     dim.3     dim.4     dim.5
#> FrenchPop.No   3.674225  2.047428 44.171667 12.200644  0.012500
#> FrenchPop.Yes  1.975134  1.247293 26.317814  8.355471  0.003221
#> Rap.No         0.229719  5.492935  0.683233  4.936589  1.025606
#> Rap.Yes        1.661392 39.217602  4.406732 33.809391  8.506834
#> Rock.No        2.024761  8.833085  3.509033 10.539278  1.619610
#> Rock.Yes       4.940688 21.861438 11.919624 29.985195  4.807727
#> Jazz.No        8.820421  0.138032  0.644780  0.000855  7.966239
#> Jazz.Yes      42.284846  0.808740  2.968706  0.000002 36.347838
#> Classical.No  10.624531  6.218211  1.742768  0.041977 12.719128
#> Classical.Yes 23.764284 14.135237  3.635642  0.130599 26.991297
#> 
#> $var.ctr
#>               dim.1     dim.2     dim.3     dim.4     dim.5
#> FrenchPop  5.649359  3.294721 70.489481 20.556115  0.015721
#> Rap        1.891111 44.710537  5.089965 38.745980  9.532440
#> Rock       6.965449 30.694523 15.428657 40.524473  6.427337
#> Jazz      51.105267  0.946772  3.613486  0.000857 44.314077
#> Classical 34.388815 20.353448  5.378410  0.172576 39.710425
#> 
#> $ctr.cloud
#>               ctr.cloud
#> FrenchPop.No      12.60
#> FrenchPop.Yes      7.80
#> Rap.No             2.76
#> Rap.Yes           17.56
#> Rock.No            5.64
#> Rock.Yes          14.76
#> Jazz.No            3.72
#> Jazz.Yes          16.56
#> Classical.No       6.48
#> Classical.Yes     13.88
#> 
#> $vctr.cloud
#>           vctr.cloud
#> FrenchPop      20.40
#> Rap            20.32
#> Rock           20.40
#> Jazz           20.28
#> Classical      20.36
#>