Computes contributions for a MCA

From MCA results, computes contributions of categories and variables to the axes and the overall cloud.

contrib(resmca)

Arguments

resmca
object of class 'MCA', 'speMCA' or 'csMCA'

Details

The contribution of a point to an axis depends both on the distance from the point to the origin point along the axis and on the weight of the point. The contributions of points to axes are the main aid to interpretation (see Le Roux and Rouanet, 2004 and 2010).

Value

A list of data frames:

ctr

Data frame with the contributions of categories to axes

var.ctr

Data frame with the contributions of variables to axes

ctr.cloud

Data frame with the contributions of categories to the overall cloud

vctr.cloud

Data frame with the contributions of variables to the overall cloud

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on the 'Music' example data set
## and compute contributions
data(Music)
mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
contrib(mca)

#> $ctr
#>                   dim.1     dim.2     dim.3     dim.4     dim.5
#> FrenchPop.No   0.516818  0.053136 53.977671  5.466075  0.801522
#> FrenchPop.Yes  0.304068  0.058209 34.853219  3.543750  0.412908
#> Rap.No         0.831162  7.556392  0.352049  6.677906  0.173217
#> Rap.Yes        4.559478 40.736581  1.920143 36.390471  0.797873
#> Rock.No        3.177022 11.186467  1.258184  8.765206  2.894161
#> Rock.Yes       8.229826 30.568889  3.076714 23.223416  7.621192
#> Jazz.No        8.513241  0.526292  0.822434  0.302381  9.118656
#> Jazz.Yes      35.344389  2.272931  2.359414  1.163931 39.589655
#> Classical.No  10.799813  2.034206  0.412300  4.352110 11.487415
#> Classical.Yes 27.724183  5.006896  0.967873 10.114755 27.103402
#> 
#> $var.ctr
#>               dim.1     dim.2     dim.3     dim.4    dim.5
#> FrenchPop  0.820886  0.111345 88.830890  9.009825  1.21443
#> Rap        5.390640 48.292973  2.272192 43.068377  0.97109
#> Rock      11.406848 41.755356  4.334898 31.988622 10.51535
#> Jazz      43.857630  2.799223  3.181848  1.466312 48.70831
#> Classical 38.523996  7.041102  1.380173 14.466865 38.59082
#> 
#> $ctr.cloud
#>               ctr.cloud
#> FrenchPop.No      12.24
#> FrenchPop.Yes      7.96
#> Rap.No             3.32
#> Rap.Yes           16.92
#> Rock.No            5.60
#> Rock.Yes          14.60
#> Jazz.No            4.20
#> Jazz.Yes          16.20
#> Classical.No       5.96
#> Classical.Yes     14.32
#> 
#> $vctr.cloud
#>           vctr.cloud
#> FrenchPop      20.20
#> Rap            20.24
#> Rock           20.20
#> Jazz           20.40
#> Classical      20.28
#>