Describes the contributions to axes for MCA and variants of MCA

Identifies the categories and individuals that contribute the most to each dimension obtained by a Multiple Correspondence Analysis. It allows to analyze variants of MCA, such as 'specific' MCA or 'class specific' MCA.

dimcontrib(resmca, dim = c(1,2), best = TRUE)

Arguments

resmca
object of class 'MCA', 'speMCA', or 'csMCA'

dim
dimensions to describe (default is c(1,2))

best
if FALSE, displays all the categories; if TRUE (default), displays only categories and individuals with contributions higher than average

Details

Contributions are sorted and assigned a positive or negative sign according to the corresponding categories or individuals' coordinates, so as to facilitate interpretation.

Value

Returns a list:

var

a list of categories' contributions to axes

ind

a list of individuals' contributions to axes

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then describes the contributions to axes.
data(Music)
getindexcat(Music[,1:5])

#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA" 

mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
dimcontrib(mca)

#> $var
#> $var$axe.1
#>                 ctr weight
#> Rock.Yes      15.86    113
#> Classical.Yes 20.73    125
#> Jazz.Yes      37.32     80
#> 
#> $var$axe.2
#>                  ctr weight
#> Classical.Yes -15.66    125
#> Rock.Yes       12.98    113
#> FrenchPop.No   14.23    173
#> Rap.Yes        37.00     47
#> 
#> 
#> $ind
#> $ind$axe.1
#>      324     2739      987     4937      235     3650     4144     2687 
#> 0.243869 0.282859 0.282859 0.312148 0.312148 0.312148 0.312148 0.312148 
#>     2635     1254     2055     4963     3745      439     3066     4546 
#> 0.312148 0.312148 0.312148 0.312148 0.429947 0.429947 0.429947 0.429947 
#>     4791     3863     4280     2259     2535     1320     1612     4718 
#> 0.499011 0.499011 0.499011 0.499011 0.499011 0.499011 0.499011 0.499011 
#>     1489     3498     2674     4119      262     4962      227      736 
#> 0.506370 0.538433 0.612902 0.612902 0.612902 0.612902 0.612902 0.612902 
#>      993     2540     3422     1859      647      188     4538     4732 
#> 0.625041 0.694862 0.694862 0.694862 0.694862 0.694862 0.694862 0.694862 
#>     3836     1456     3754     3447      436     4651     3119     4211 
#> 0.694862 0.694862 0.694862 0.694862 0.694862 0.694862 0.694862 0.694862 
#>      920      707     3345     3931     2374     3304     3208     2456 
#> 0.875676 1.129868 1.129868 1.129868 1.129868 1.232851 1.232851 1.232851 
#>     2957     3244     2033     1418        4     2843     4495     2132 
#> 1.232851 1.531443 1.531443 1.531443 1.531443 1.531443 1.531443 1.531443 
#>     4834     4477     2471     3312     2364     1469     3556     4852 
#> 1.531443 1.531443 1.599941 1.933240 1.933240 2.303070 2.303070 2.303070 
#>     1163 
#> 2.303070 
#> 
#> $ind$axe.2
#>      2540      3422      1859       647       188      4538      4732      3836 
#> -0.431896 -0.431896 -0.431896 -0.431896 -0.431896 -0.431896 -0.431896 -0.431896 
#>      1456      3754      3447       436      4651      3119      4211      2169 
#> -0.431896 -0.431896 -0.431896 -0.431896 -0.431896 -0.431896 -0.431896 -0.412042 
#>       280      1206      2438      1392      2590       632        58      2592 
#> -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 
#>      2720      1226      4245      4526       108      1486      4240      2090 
#> -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 
#>      1407      2110      3819      3023      1246      4027       340      2673 
#> -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 
#>      2601       737      4115      2740       871      4398      1160        42 
#> -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 
#>      1960      4035      2137      2621       272      1352      1300      3081 
#> -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 
#>      4303      2143      1311       236      3512      2040      2827      2499 
#> -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 -0.412042 
#>       908      4747      1107      4341      4635      1602       993      3045 
#> -0.412042 -0.412042 -0.412042 -0.412042 -0.302653 -0.302653 -0.268531  0.242235 
#>      4759      3745       439      3066      4546      4443      4694       823 
#>  0.242235  0.342507  0.342507  0.342507  0.342507  0.360629  0.360629  0.360629 
#>       194      1790      1707      4812      1614      4366      2488      2947 
#>  0.360629  0.360629  0.360629  0.360629  0.360629  0.360629  0.360629  0.360629 
#>       597      1782      4498      1314      4746      1084      3715      1850 
#>  0.360629  0.360629  0.360629  0.360629  0.360629  0.360629  0.360629  0.360629 
#>      2069      1389      3272      1469      3556      4852      1163      2408 
#>  0.360629  0.360629  0.360629  0.423905  0.423905  0.423905  0.423905  0.442923 
#>       804      2858      4355       776      3950      4327      2914      3648 
#>  0.485518  0.485518  0.485518  0.485518  0.485518  0.485518  0.485518  0.485518 
#>      4720      4493      3978       896      3312      2364      1489      2739 
#>  0.485518  0.485518  0.485518  0.485518  1.112015  1.112015  1.144481  1.177116 
#>       987       707      3345      3931      2374      4472      4143      3664 
#>  1.177116  1.178936  1.178936  1.178936  1.178936  1.210511  1.210511  1.210511 
#>      2043       895      1560      2578      2848      2661      2218      4989 
#>  1.210511  1.210511  1.210511  1.210511  1.210511  1.210511  1.210511  1.210511 
#>      1653      3656      2447      1310      4170      1360      3264       223 
#>  1.210511  1.212357  1.212357  1.212357  1.212357  1.212357  1.212357  1.212357 
#>       920      2594 
#>  2.217804  2.263556 
#> 
#>