Describes the contributions to axes for a MCA

Identifies the categories and individuals that contribute the most to each dimension obtained by a Multiple Correspondence Analysis. It allows to analyze variants of MCA, such as 'specific' MCA or 'class specific' MCA.

dimcontrib(resmca, dim = c(1,2), best = TRUE)

Arguments

resmca

object of class 'MCA', 'speMCA', or 'csMCA'

dim

dimensions to describe (default is c(1,2))

best

if FALSE, displays all the categories; if TRUE (default), displays only categories and individuals with contributions higher than average

Details

Contributions are sorted and assigned a positive or negative sign according to the corresponding categories or individuals' coordinates, so as to facilitate interpretation.

Value

Returns a list:

var

a list of categories' contributions to axes

ind

a list of individuals' contributions to axes

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then describes the contributions to axes.
data(Music)
getindexcat(Music[,1:5])
#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA" 
mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
dimcontrib(mca)
#> $var
#> $var$dim.1
#>                  ctr weight
#> Jazz.Yes      -35.34     95
#> Classical.Yes -27.72    142
#> Classical.No   10.80    351
#> 
#> $var$dim.2
#>             ctr weight
#> Rap.Yes  -40.74     77
#> Rock.Yes -30.57    135
#> Rock.No   11.19    360
#> 
#> 
#> $ind
#> $ind$dim.1
#>      4900      4763       986      1419      1762      2017      1624      4021 
#> -1.357604 -1.357604 -1.357604 -1.357604 -1.357604 -1.357604 -1.357604 -1.357604 
#>      3857      1642      4064      3208      2660      2957      3694      3413 
#> -1.357604 -1.357604 -1.197912 -1.197912 -1.197912 -1.197912 -1.197912 -1.197912 
#>      4491      2004      1147      1049      3574      4308        11      1982 
#> -0.913815 -0.848472 -0.848472 -0.848472 -0.767956 -0.767956 -0.767956 -0.767956 
#>      3775      3767      3109       643       516       621      1362      2841 
#> -0.767956 -0.767956 -0.767956 -0.767956 -0.767956 -0.767956 -0.767956 -0.767956 
#>      3557      1023      2106      4990      2364      1387      1291      3863 
#> -0.767956 -0.767956 -0.767956 -0.723274 -0.723274 -0.721477 -0.649088 -0.649088 
#>      1627      1704        23      4046      1185      3788      3340      2535 
#> -0.649088 -0.649088 -0.649088 -0.649088 -0.649088 -0.649088 -0.649088 -0.649088 
#>      4984      2259      3436      2675      1857      2394      4845      1481 
#> -0.649088 -0.649088 -0.649088 -0.649088 -0.649088 -0.649088 -0.412279 -0.412279 
#>      1924      2508      4304      4661      1212      4707       460       865 
#> -0.412279 -0.412279 -0.412279 -0.412279 -0.412279 -0.412279 -0.412279 -0.412279 
#>      1011       893      2828      2489      3636      2066      1089       506 
#> -0.412279 -0.333642 -0.333304 -0.326520 -0.326520 -0.326520 -0.326520 -0.274455 
#>      1459      3639        20      2950      3858       304       689      2441 
#> -0.274455 -0.274455 -0.274455 -0.274455 -0.274455 -0.274455 -0.274455 -0.274455 
#>      4755       296      3623      1899      2934      1553       297      4838 
#> -0.274455 -0.205402 -0.205402 -0.205402 -0.205402 -0.205402 -0.205402  0.206194 
#>       459      4485      4742      3386      4010      4187      1327      3621 
#>  0.206194  0.228793  0.228793  0.283084  0.283084  0.283084  0.283084  0.283084 
#>      2100      1501      2500      1436      4563      1834      4698      1328 
#>  0.283084  0.283084  0.283084  0.283084  0.283084  0.283084  0.283084  0.283084 
#>      2750      1238      1016      4154      3094       620      4167      1714 
#>  0.283084  0.283084  0.283084  0.283084  0.283084  0.283084  0.283084  0.283084 
#>       450      2218       377      3777       421      2421      2578      1202 
#>  0.283084  0.312470  0.363280  0.363280  0.363280  0.363280  0.363280  0.363280 
#>      4668      3726      2043      4151      2034       381      3294      4565 
#>  0.363280  0.363280  0.363280  0.363280  0.363280  0.363280  0.363280  0.363280 
#>      1040      1441      2285       963      1313      4361 
#>  0.363280  0.363280  0.363280  0.363280  0.363280  0.363280 
#> 
#> $ind$dim.2
#>      4757      4564       740      1848      4705      3550      3478      3760 
#> -1.930377 -1.930377 -1.850166 -1.850166 -1.850166 -1.455065 -1.455065 -1.385538 
#>      2961      2419      4312       223      4076      1882      1986       420 
#> -1.385538 -1.385538 -1.385538 -1.385538 -1.385538 -1.385538 -1.385538 -1.385538 
#>      1094      4990      2364      2004      1147      1049      1489      1665 
#> -1.385538 -1.291869 -1.291869 -1.226406 -1.226406 -1.226406 -0.909133 -0.854354 
#>      1972      1250      3452      2828      4485      4742       377      3777 
#> -0.854354 -0.541072 -0.541072 -0.458115 -0.380185 -0.380185 -0.338292 -0.338292 
#>       421      2421      2578      1202      4668      3726      2043      4151 
#> -0.338292 -0.338292 -0.338292 -0.338292 -0.338292 -0.338292 -0.338292 -0.338292 
#>      2034       381      3294      4565      1040      1441      2285       963 
#> -0.338292 -0.338292 -0.338292 -0.338292 -0.338292 -0.338292 -0.338292 -0.338292 
#>      1313      4361      2218      2489      3636      2066      1089      3386 
#> -0.338292 -0.338292 -0.320103 -0.317821 -0.317821 -0.317821 -0.317821 -0.305208 
#>      4010      4187      1327      3621      2100      1501      2500      1436 
#> -0.305208 -0.305208 -0.305208 -0.305208 -0.305208 -0.305208 -0.305208 -0.305208 
#>      4563      1834      4698      1328      2750      1238      1016      4154 
#> -0.305208 -0.305208 -0.305208 -0.305208 -0.305208 -0.305208 -0.305208 -0.305208 
#>      3094       620      4167      1714       450      4845      1481      1924 
#> -0.305208 -0.305208 -0.305208 -0.305208 -0.305208 -0.285781 -0.285781 -0.285781 
#>      2508      4304      4661      1212      4707       460       865      1011 
#> -0.285781 -0.285781 -0.285781 -0.285781 -0.285781 -0.285781 -0.285781 -0.285781 
#>       371      4838       459      4246      4618       641      3508      1546 
#> -0.259666 -0.258787 -0.258787  0.213267  0.213267  0.241062  0.246782  0.246782 
#>      2846       151      3126      1945      1093      1176      1075      4043 
#>  0.246782  0.246782  0.246782  0.246782  0.246782  0.246782  0.246782  0.246782 
#>      4107       949      2026      3942      1820      3986      1932      1670 
#>  0.246782  0.246782  0.246782  0.246782  0.246782  0.276616  0.276616  0.276616 
#>      2404        31      4274       280      3793       701      1959      3769 
#>  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616 
#>      4814      2114       984      3286        58      3702      4000      2827 
#>  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616 
#>       161      1111      4367       696        88      4100      2312      2610 
#>  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616 
#>       337      1059      1506      3377      1433        89      2906      4930 
#>  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616 
#>       422       106       180      3657       927      4458      4251      1877 
#>  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616  0.276616 
#>      3063 
#>  0.276616 
#> 
#>