Describes the dimensions of a MCA

Identifies the variables and the categories that are the most characteristic according to each dimension obtained by a MCA. It is inspired by dimdesc function in FactoMineR package (see Husson et al, 2010), but allows to analyze variants of MCA, such as 'specific' MCA or 'class specific' MCA.

dimdescr(resmca, vars=NULL, dim = c(1,2), min.cor = NULL, 
nperm = 100, distrib = "asympt")

Arguments

resmca

object of class MCA, speMCA, csMCA, stMCA or multiMCA

vars

data frame of variables to describes the MCA dimensions with. If NULL (default), the active variables of the MCA will be used.

dim

the axes which are described. Default is c(1,2)

min.cor

for the relationship between y and a categorical variable, only associations higher or equal to min.cor will be displayed. If NULL (default), they are all displayed.

nperm

numeric. Number of permutations for the permutation test of independence. If NULL, no permutation test is performed.

distrib

the null distribution of permutation test of independence can be approximated by its asymptotic distribution ("asympt", default) or via Monte Carlo resampling ("approx").

Details

See condesc.

Value

Returns a list of ncp lists including:

variables

associations between y and the variables in x

categories

a data frame with categorical variables from x and associations measured by correlation coefficients

References

Husson, F., Le, S. and Pages, J. (2010). Exploratory Multivariate Analysis by Example Using R, Chapman and Hall.

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then describe the dimensions.
data(Music)
getindexcat(Music[,1:5])
#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA" 
mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
dimdescr(mca,min.cor=0.1,nperm=10)
#> $dim.1
#> $dim.1$variables
#>    variable measure association permutation.pvalue
#> 1      Jazz    Eta2       0.611                  0
#> 2 Classical    Eta2       0.539                  0
#> 3      Rock    Eta2       0.160                  0
#> 4       Rap    Eta2       0.075                  0
#> 5 FrenchPop    Eta2       0.012                  0
#> 
#> $dim.1$categories
#>       categories mean.y.in.xcat mean.y.global sd.y.in.xcat sd.y.global    cor
#> 10       Jazz.No          0.204             0        0.324       0.528  0.751
#> 13  Classical.No          0.244             0        0.335       0.528  0.710
#> 7        Rock.No          0.131             0        0.471       0.528  0.398
#> 5        Rap.Yes          0.339             0        0.504       0.528  0.274
#> 1   FrenchPop.No          0.072             0        0.543       0.528  0.108
#> 2  FrenchPop.Yes         -0.044             0        0.511       0.528 -0.103
#> 4         Rap.No         -0.062             0        0.510       0.528 -0.264
#> 8       Rock.Yes         -0.344             0        0.514       0.528 -0.396
#> 14 Classical.Yes         -0.615             0        0.414       0.528 -0.734
#> 11      Jazz.Yes         -0.849             0        0.340       0.528 -0.780
#> 
#> 
#> $dim.2
#> $dim.2$variables
#>    variable measure association permutation.pvalue
#> 1       Rap    Eta2       0.526              0.000
#> 2      Rock    Eta2       0.456              0.000
#> 3 Classical    Eta2       0.077              0.000
#> 4      Jazz    Eta2       0.031              0.000
#> 5 FrenchPop    Eta2       0.003              0.352
#> 
#> $dim.2$categories
#>       categories mean.y.in.xcat mean.y.global sd.y.in.xcat sd.y.global    cor
#> 4         Rap.No          0.147             0        0.320       0.467  0.704
#> 7        Rock.No          0.192             0        0.339       0.467  0.660
#> 14 Classical.Yes          0.205             0        0.445       0.467  0.276
#> 10       Jazz.No          0.040             0        0.443       0.467  0.165
#> 11      Jazz.Yes         -0.169             0        0.526       0.467 -0.175
#> 13  Classical.No         -0.083             0        0.451       0.467 -0.273
#> 8       Rock.Yes         -0.519             0        0.363       0.467 -0.676
#> 5        Rap.Yes         -0.793             0        0.318       0.467 -0.724
#> 
#>