Describes the dimensions of MCA and variants of MCA

Identifies the variables and the categories that are the most characteristic according to each dimension obtained by a MCA. It is inspired by dimdesc function in FactoMineR package (see Husson et al, 2010), but allows to analyze variants of MCA, such as 'specific' MCA or 'class specific' MCA.

dimdescr(resmca, ncp = 1:3, min.cor = NULL, nperm = 100, distrib = "asympt")

Arguments

resmca
object of class 'MCA', 'speMCA', 'csMCA' or 'stMCA'

ncp
number of dimensions to describe (default is 3)

min.cor
for the relationship between y and a categorical variable, only associations higher or equal to min.cor will be displayed. If NULL (default), they are all displayed.

nperm
numeric. Number of permutations for the permutation test of independence. If NULL, no permutation test is performed.

distrib
the null distribution of permutation test of independence can be approximated by its asymptotic distribution ("asympt", default) or via Monte Carlo resampling ("approx").

Details

See condesc.

Value

Returns a list of ncp lists including:

variables

associations between y and the variables in x

categories

a data frame with categorical variables from x and associations measured by correlation coefficients

References

Husson, F., Le, S. and Pages, J. (2010). Exploratory Multivariate Analysis by Example Using R, Chapman and Hall.

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then describe the dimensions.
data(Music)
getindexcat(Music[,1:5])

#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA" 

mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
dimdescr(mca,min.cor=0.1,nperm=10)

#> [1] 1
#> [1] 2
#> [1] 3
#> [1] 4
#> [1] 5
#> [1] 6
#> [1] 7
#> [1] 8
#> [1] 9
#> [1] 10
#> [1] 11
#> [1] 12
#> [1] 13
#> [1] 14
#> [1] 15
#> [1] 1
#> [1] 2
#> [1] 3
#> [1] 4
#> [1] 5
#> [1] 6
#> [1] 7
#> [1] 8
#> [1] 9
#> [1] 10
#> [1] 11
#> [1] 12
#> [1] 13
#> [1] 14
#> [1] 15
#> [1] 1
#> [1] 2
#> [1] 3
#> [1] 4
#> [1] 5
#> [1] 6
#> [1] 7
#> [1] 8
#> [1] 9
#> [1] 10
#> [1] 11
#> [1] 12
#> [1] 13
#> [1] 14
#> [1] 15

#> $dim.1
#> $dim.1$variables
#>    variable measure association permutation.pvalue
#> 1      Jazz    Eta2       0.639                  0
#> 2 Classical    Eta2       0.399                  0
#> 3      Rock    Eta2       0.294                  0
#> 4       Rap    Eta2       0.069                  0
#> 5 FrenchPop    Eta2       0.044                  0
#> 
#> $dim.1$categories
#>       categories median.y.in.xcat median.y.global mad.y.in.xcat mad.y.global
#> 11      Jazz.Yes       0.95325220      -0.1757057     0.3144685     0.299283
#> 14 Classical.Yes       0.38076161      -0.1757057     0.3129657     0.299283
#> 8       Rock.Yes       0.33221044      -0.1757057     0.3354517     0.299283
#> 5        Rap.Yes       0.22747448      -0.1757057     0.3600842     0.299283
#> 2  FrenchPop.Yes      -0.02391302      -0.1757057     0.3275387     0.299283
#> 1   FrenchPop.No      -0.35207123      -0.1757057     0.2005251     0.299283
#> 4         Rap.No      -0.25035395      -0.1757057     0.2632684     0.299283
#> 7        Rock.No      -0.31777536      -0.1757057     0.1722030     0.299283
#> 13  Classical.No      -0.33752004      -0.1757057     0.1448940     0.299283
#> 10       Jazz.No      -0.30741385      -0.1757057     0.1960495     0.299283
#>      cor.coef
#> 11  0.7990212
#> 14  0.6302762
#> 8   0.5426220
#> 5   0.2615363
#> 2   0.1890679
#> 1  -0.1967560
#> 4  -0.2501754
#> 7  -0.5141834
#> 13 -0.6227395
#> 10 -0.7852388
#> 
#> 
#> $dim.2
#> $dim.2$variables
#>    variable measure association permutation.pvalue
#> 1       Rap    Eta2       0.481              0.000
#> 2 FrenchPop    Eta2       0.257              0.000
#> 3 Classical    Eta2       0.246              0.000
#> 4      Rock    Eta2       0.197              0.000
#> 5      Jazz    Eta2       0.000              0.896
#> 
#> $dim.2$categories
#>       categories median.y.in.xcat median.y.global mad.y.in.xcat mad.y.global
#> 5        Rap.Yes        1.1195488      -0.1360857     0.1261472    0.2991082
#> 1   FrenchPop.No        0.2421012      -0.1360857     0.1174346    0.2991082
#> 13  Classical.No        0.1229966      -0.1360857     0.3189994    0.2991082
#> 8       Rock.Yes        0.2998116      -0.1360857     0.3107064    0.2991082
#> 7        Rock.No       -0.1982059      -0.1360857     0.2588778    0.2991082
#> 14 Classical.Yes       -0.5770787      -0.1360857     0.1742586    0.2991082
#> 2  FrenchPop.Yes       -0.2260146      -0.1360857     0.1091604    0.2991082
#> 4         Rap.No       -0.1787894      -0.1360857     0.3249515    0.2991082
#>      cor.coef
#> 5   0.6930862
#> 1   0.5059003
#> 13  0.4807570
#> 8   0.4441386
#> 7  -0.4225395
#> 14 -0.4955634
#> 2  -0.5003524
#> 4  -0.6423266
#> 
#> 
#> $dim.3
#> $dim.3$variables
#>    variable measure association permutation.pvalue
#> 1 FrenchPop    Eta2       0.694                  0
#> 2 Classical    Eta2       0.090                  0
#> 3      Rock    Eta2       0.065                  0
#> 4       Rap    Eta2       0.058                  0
#> 5      Jazz    Eta2       0.054                  0
#> 
#> $dim.3$categories
#>       categories median.y.in.xcat median.y.global mad.y.in.xcat mad.y.global
#> 2  FrenchPop.Yes       0.28876376      0.07201885     0.0912397    0.2926809
#> 13  Classical.No       0.22112468      0.07201885     0.3068639    0.2926809
#> 8       Rock.Yes       0.28159291      0.07201885     0.3383765    0.2926809
#> 5        Rap.Yes       0.41270724      0.07201885     0.4544296    0.2926809
#> 10       Jazz.No       0.14875569      0.07201885     0.2456526    0.2926809
#> 4         Rap.No       0.05625130      0.07201885     0.2871521    0.2926809
#> 11      Jazz.Yes      -0.16416542      0.07201885     0.2734894    0.2926809
#> 7        Rock.No       0.02803382      0.07201885     0.2969357    0.2926809
#> 14 Classical.Yes      -0.04726666      0.07201885     0.1568487    0.2926809
#> 1   FrenchPop.No      -0.50015700      0.07201885     0.0595626    0.2926809
#>      cor.coef
#> 2   0.8328657
#> 13  0.2979073
#> 8   0.2550064
#> 5   0.2344694
#> 10  0.2088168
#> 4  -0.1928332
#> 11 -0.2252434
#> 7  -0.2424768
#> 14 -0.2822321
#> 1  -0.8285097
#> 
#>