Describes the eta2 of a list of supplementary variables for the axes of MCA and variants of MCA

Computes eta2 for a list of supplementary variables. It allows to analyze variants of MCA, such as 'specific' MCA or 'class specific' MCA.

dimeta2(resmca, vars, dim = 1:resmca$call$ncp)

Arguments

resmca
object of class 'MCA', 'speMCA', 'csMCA' or 'stMCA'

vars
a data frame of supplementary variables

dim
the axes for which eta2 are computed. Default is 1:ncp

Value

Returns a data frame with supplementary variables as rows and axes as columns.

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then describes the eta2 for Gender and Age (axes 1 and 2).
data(Music)
getindexcat(Music[,1:5])

#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA" 

mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
dimeta2(mca, Music[,c("Gender","Age")])

#>        dim.1 dim.2 dim.3 dim.4 dim.5
#> Gender   0.0   0.9   0.2   0.0   0.0
#> Age      3.8  22.7   2.0   0.8   0.2