Performs a 'specific' MCA

Performs a 'specific' Multiple Correspondence Analysis, i.e. a variant of MCA that allows to treat undesirable categories as passive categories.

speMCA(data, excl = NULL, ncp = 5, row.w = rep(1, times = nrow(data)))

Arguments

data
data frame with n rows (individuals) and p columns (categorical variables)

excl
numeric vector indicating the indexes of the "junk" categories (default is NULL). See "getindexcat" to identify these indexes.

ncp
number of dimensions kept in the results (default is 5)

row.w
an optional numeric vector of row weights (by default, a vector of 1 for uniform row weights)

Details

Undesirable categories may be of several kinds: infrequent categories (say, <5 percents), heterogeneous categories (e.g. 'others') or uninterpretable categories (e.g. 'not available'). In these cases, 'specific' MCA may be useful to ignore these categories for the determination of distances between individuals (see Le Roux and Rouanet, 2004 and 2010).

Value

Returns an object of class 'speMCA', i.e. a list including:

eig

a list of vectors containing all the eigenvalues, the percentage of variance, the cumulative percentage of variance, the modified rates and the cumulative modified rates

call

a list with informations about input data

ind

a list of matrices containing the results for the individuals (coordinates, contributions)

var

a list of matrices containing all the results for the categories and variables (weights, coordinates, square cosine, categories contributions to axes and cloud, test values (v.test), square correlation ratio (eta2), variable contributions to axes and cloud

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories.
data(Music)
getindexcat(Music[,1:5])

#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA" 

mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
str(mca)

#> List of 5
#>  $ eig :List of 5
#>   ..$ eigen    : num [1:10] 0.264 0.237 0.201 0.172 0.125 ...
#>   ..$ rate     : num [1:10] 26 23.3 19.8 16.9 12.3 ...
#>   ..$ cum.rate : num [1:10] 26 49.3 69.1 86 98.3 ...
#>   ..$ mrate    : num [1:3] 74.59 25.38 0.03
#>   ..$ cum.mrate: num [1:3] 74.6 100 100
#>  $ call:List of 7
#>   ..$ X        :'data.frame':	500 obs. of  5 variables:
#>   .. ..$ FrenchPop: Factor w/ 3 levels "No","Yes","NA": 2 2 2 2 2 3 2 2 1 2 ...
#>   .. ..$ Rap      : Factor w/ 3 levels "No","Yes","NA": 1 2 1 1 1 1 1 1 1 1 ...
#>   .. ..$ Rock     : Factor w/ 3 levels "No","Yes","NA": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ Jazz     : Factor w/ 3 levels "No","Yes","NA": 1 1 1 1 1 2 2 1 1 1 ...
#>   .. ..$ Classical: Factor w/ 3 levels "No","Yes","NA": 1 2 1 2 1 2 2 2 1 2 ...
#>   ..$ marge.col: Named num [1:10] 0.074 0.122 0.1724 0.0244 0.1436 ...
#>   .. ..- attr(*, "names")= chr [1:10] "FrenchPop.No" "FrenchPop.Yes" "Rap.No" "Rap.Yes" ...
#>   ..$ marge.row: Named num [1:500] 4e-04 4e-04 4e-04 4e-04 4e-04 4e-04 4e-04 4e-04 4e-04 4e-04 ...
#>   .. ..- attr(*, "names")= chr [1:500] "1" "2" "3" "4" ...
#>   ..$ ncp      : num 5
#>   ..$ quali    : int [1:5] 1 2 3 4 5
#>   ..$ excl     : num [1:5] 3 6 9 12 15
#>   ..$ row.w    : num [1:500] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ ind :List of 2
#>   ..$ coord  : num [1:500, 1:5] 0.338 -0.422 0.338 -0.234 0.338 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:500] "3324" "3122" "1597" "4589" ...
#>   .. .. ..$ : chr [1:5] "dim.1" "dim.2" "dim.3" "dim.4" ...
#>   ..$ contrib: num [1:500, 1:5] 0.0865 0.1346 0.0865 0.0413 0.0865 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:500] "3324" "3122" "1597" "4589" ...
#>   .. .. ..$ : chr [1:5] "dim.1" "dim.2" "dim.3" "dim.4" ...
#>  $ var :List of 9
#>   ..$ weight    : Named num [1:10] 185 305 431 61 359 131 407 86 338 153
#>   .. ..- attr(*, "names")= chr [1:10] "FrenchPop.No" "FrenchPop.Yes" "Rap.No" "Rap.Yes" ...
#>   ..$ coord     : num [1:10, 1:5] 0.3621 -0.2068 0.0593 -0.424 0.193 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:10] "FrenchPop.No" "FrenchPop.Yes" "Rap.No" "Rap.Yes" ...
#>   .. .. ..$ : chr [1:5] "dim.1" "dim.2" "dim.3" "dim.4" ...
#>   ..$ contrib   : num [1:10, 1:5] 3.67 1.98 0.23 1.66 2.02 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:10] "FrenchPop.No" "FrenchPop.Yes" "Rap.No" "Rap.Yes" ...
#>   .. .. ..$ : chr [1:5] "dim.1" "dim.2" "dim.3" "dim.4" ...
#>   ..$ ctr.cloud :'data.frame':	10 obs. of  1 variable:
#>   .. ..$ ctr.cloud: num [1:10] 12.6 7.8 2.76 17.56 5.64 ...
#>   ..$ cos2      : num [1:10, 1:5] 0.077 0.0669 0.022 0.025 0.0948 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:10] "FrenchPop.No" "FrenchPop.Yes" "Rap.No" "Rap.Yes" ...
#>   .. .. ..$ : chr [1:5] "dim.1" "dim.2" "dim.3" "dim.4" ...
#>   ..$ v.test    : num [1:10, 1:5] 6.2 -5.78 3.31 -3.53 6.88 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:10] "FrenchPop.No" "FrenchPop.Yes" "Rap.No" "Rap.Yes" ...
#>   .. .. ..$ : chr [1:5] "dim.1" "dim.2" "dim.3" "dim.4" ...
#>   ..$ eta2      : num [1:5, 1:5] 0.0777 0.025 0.095 0.6751 0.4543 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:5] "FrenchPop" "Rap" "Rock" "Jazz" ...
#>   .. .. ..$ : chr [1:5] "dim.1" "dim.2" "dim.3" "dim.4" ...
#>   ..$ v.contrib :'data.frame':	5 obs. of  5 variables:
#>   .. ..$ dim.1: num [1:5] 5.65 1.89 6.97 51.11 34.39
#>   .. ..$ dim.2: num [1:5] 3.295 44.711 30.695 0.947 20.353
#>   .. ..$ dim.3: num [1:5] 70.49 5.09 15.43 3.61 5.38
#>   .. ..$ dim.4: num [1:5] 2.06e+01 3.87e+01 4.05e+01 8.58e-04 1.73e-01
#>   .. ..$ dim.5: num [1:5] 0.0157 9.5324 6.4273 44.3141 39.7104
#>   ..$ vctr.cloud:'data.frame':	5 obs. of  1 variable:
#>   .. ..$ vctr.cloud: num [1:5] 20.4 20.3 20.4 20.3 20.4
#>  $ svd :List of 3
#>   ..$ vs: num [1:10] 0.514 0.487 0.449 0.415 0.354 ...
#>   ..$ U : num [1:500, 1:10] 0.0294 -0.0367 0.0294 -0.0203 0.0294 ...
#>   ..$ V : num [1:10, 1:10] 0.1917 -0.1405 0.0479 -0.1289 0.1423 ...
#>  - attr(*, "class")= chr [1:2] "speMCA" "list"