Displays the categories contributing most to axes for a MCA

Identifies the categories that contribute the most to a given dimension of a Multiple Correspondence Analysis and organizes these informations into a fancy table. It allows to analyze variants of MCA, such as 'specific' MCA or 'class specific' MCA.

tabcontrib(resmca, dim = 1, best = TRUE)

Arguments

resmca

object of class MCA, speMCA, or csMCA

dim

dimension to describe (default is 1st dimension)

best

if FALSE, displays all the categories; if TRUE (default), displays only categories and individuals with contributions higher than average

Details

Best contributions - i.e. higher than average - are assigned a positive or negative sign according to the corresponding categories' coordinates, so as to facilitate interpretation. Then they are sorted and organized according to the most contributing variables.

Value

Returns a data frame with the following columns:

var

the names of the most contributing variables

moda

the names of the most contributing cateories

ctr1

'negative' contributions, i.e. corresponding to categories with coordinates lower than zero

ctr2

'positive' contributions, i.e. corresponding to categories with coordinates higher than zero

weight

weight of the categories

ctrtot

sum of the best contributions for a given variable

cumctrtot

cumulated contributions

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then describes the contributions to axes.
data(Music)
getindexcat(Music[,1:5])
#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA" 
mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
tabcontrib(mca,1)
#>         var moda   ctr1 ctr2 weight ctrtot cumctrtot
#> 1 Classical  Yes -27.72         142  38.52     38.52
#> 2             No        10.8    351                 
#> 3      Jazz  Yes -35.34          95  35.34     73.86
tabcontrib(mca,2)
#>    var moda   ctr1  ctr2 weight ctrtot cumctrtot
#> 2 Rock  Yes -30.57          135  41.76     41.76
#> 3        No        11.19    360                 
#> 1  Rap  Yes -40.74           77  40.74      82.5