Translates logit regression coefficients into percentages

Performs a logit regression and then computes the effects of covariates expressed in percentages (through two methods: 'pure' effects and 'experimental' effects; see Deauvieau, 2010)

translate.logit(formula,data,nit=0)

Arguments

formula

an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. Every variables have to be factors.

data

a data frame containing the variables in the model

nit

number of bootstrap iterations for confidence interval computation. Default is 0, i.e. no confidence interval is computed.

Details

This function works with binomial as well as multinomial regression models. If the dependant variable has two factors, glm is used, if it has more than two factors multinom function (from nnet package) is used. The function expresses the regression coefficients as percentages through three distinct methods: raw percentages, 'pure effects' percentages and 'experimental effects' percentages (see Deauvieau, 2010). Bootstrap confidence interval are available only for binomial regressions.

Value

The function returns a list:

glm

An object of class glm or nnet (depending on the number of factors of the dependent variable)

summary

The results of summary function applied to reg element

percents

A matrix or a list of matrices (depending on the number of factors of the dependent variable) with regression coefficients expressed as percentages

boot.ci

A matrix or a list of matrices (depending on the number of factors of the dependent variable) with confidence intervals computed with bootstrap

References

Deauvieau, J. (2010), 'Comment traduire sous forme de probabilites les resultats d'une modelisation logit ?', Bulletin of Sociological Methodology / Bulletin de Methodologie Sociologique 105(1), 5-23.

Deauvieau, J. (2011), 'Est-il possible et souhaitable traduire sous forme de probabilites un coefficient logit ? Reponse aux remarques formulees par Marion Selz a propos de mon article paru dans le BMS en 2010', Bulletin of Sociological Methodology / Bulletin de Methodologie Sociologique 112(1), 32-42.

Deauvieau, J. (2019), 'Comparer les resultats d’un modele logit dichotomique ou polytomique entre plusieurs groupes a partir des probabilites estimees', Bulletin of Sociological Methodology / Bulletin de Methodologie Sociologique 142(1), 7-31.

Author

Nicolas Robette

See also

Examples

  ## An example for binomial logit regression
  data(Music)
  translate.logit(Daily ~ Gender + Age, Music)
#> $glm
#> 
#> Call:  glm(formula = f, family = "binomial", data = d)
#> 
#> Coefficients:
#> (Intercept)  GenderWomen     Age25-49       Age50+  
#>     1.22058     -0.03483     -1.20265     -2.85804  
#> 
#> Degrees of Freedom: 499 Total (i.e. Null);  496 Residual
#> Null Deviance:	    670.5 
#> Residual Deviance: 559.1 	AIC: 567.1
#> 
#> $summary
#> 
#> Call:
#> glm(formula = f, family = "binomial", data = d)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.7200  -0.5962  -0.5867   1.1698   1.9206  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  1.22058    0.28633   4.263 2.02e-05 ***
#> GenderWomen -0.03483    0.20619  -0.169    0.866    
#> Age25-49    -1.20265    0.30313  -3.967 7.27e-05 ***
#> Age50+      -2.85804    0.32598  -8.768  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 670.50  on 499  degrees of freedom
#> Residual deviance: 559.14  on 496  degrees of freedom
#> AIC: 567.14
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#> 
#> $percents
#>                   raw     expe     pure expe.lin
#> Gender.Men   0.395257 0.397271 0.398108 0.397242
#> Gender.Women 0.392713 0.390708 0.389792 0.390679
#> Age.15-24    0.769231 0.769109 0.769100 0.769102
#> Age.25-49    0.500000 0.500180 0.500146 0.500136
#> Age.50+      0.160550 0.160493 0.160460 0.160469
#> 
  
  ## An example for multinomial logit regression
  translate.logit(OnlyMus ~ Gender + Age, Music)
#> # weights:  20 (12 variable)
#> initial  value 693.147181 
#> iter  10 value 575.491808
#> iter  20 value 566.988401
#> iter  20 value 566.988401
#> iter  20 value 566.988401
#> final  value 566.988401 
#> converged
#> $glm
#> Call:
#> multinom(formula = formula, data = data, model = TRUE)
#> 
#> Coefficients:
#>       (Intercept) GenderWomen  Age25-49    Age50+
#> Often   0.5328649  -0.5309519 0.9787264 0.3129518
#> Rare   -0.5591623  -0.2590910 1.3391744 0.7967913
#> Never  -0.1497405  -0.3320937 2.0772977 2.3483931
#> 
#> Residual Deviance: 1133.977 
#> AIC: 1157.977 
#> 
#> $summary
#> Call:
#> multinom(formula = formula, data = data, model = TRUE)
#> 
#> Coefficients:
#>       (Intercept) GenderWomen  Age25-49    Age50+
#> Often   0.5328649  -0.5309519 0.9787264 0.3129518
#> Rare   -0.5591623  -0.2590910 1.3391744 0.7967913
#> Never  -0.1497405  -0.3320937 2.0772977 2.3483931
#> 
#> Std. Errors:
#>       (Intercept) GenderWomen  Age25-49    Age50+
#> Often   0.3274038   0.3255352 0.3972062 0.4042273
#> Rare    0.4156099   0.3695053 0.4766832 0.4868153
#> Never   0.3651359   0.3054313 0.4219343 0.4100501
#> 
#> Residual Deviance: 1133.977 
#> AIC: 1157.977 
#> 
#> $percents
#> $percents$by_category
#> $percents$by_category$Daily
#>                    raw      expe      pure  expe.lin
#> Gender.Men   -1.120542 -1.126346 -1.134166 -1.126143
#> Gender.Women -0.914440 -0.910035 -0.902259 -0.909877
#> Age.15-24    -0.323013 -0.321296 -0.316503 -0.320395
#> Age.25-49    -1.277271 -1.279643 -1.281633 -1.279823
#> Age.50+      -1.049945 -1.049973 -1.049768 -1.049764
#> 
#> $percents$by_category$Often
#>                   raw     expe     pure expe.lin
#> Gender.Men   0.728261 0.728822 0.730833 0.728896
#> Gender.Women 0.620690 0.619934 0.617969 0.620018
#> Age.15-24    0.568627 0.569154 0.567698 0.569081
#> Age.25-49    0.773333 0.773990 0.775360 0.774128
#> Age.50+      0.641509 0.639807 0.639537 0.639949
#> 
#> $percents$by_category$Rare
#>                   raw     expe     pure expe.lin
#> Gender.Men   0.553571 0.561332 0.563931 0.561224
#> Gender.Women 0.500000 0.493587 0.491210 0.493507
#> Age.15-24    0.333333 0.329701 0.328690 0.329532
#> Age.25-49    0.653061 0.653721 0.654244 0.653741
#> Age.50+      0.525000 0.527472 0.527382 0.527303
#> 
#> $percents$by_category$Never
#>                   raw     expe     pure expe.lin
#> Gender.Men   0.838710 0.836192 0.839402 0.836023
#> Gender.Women 0.793750 0.796514 0.793080 0.796352
#> Age.15-24    0.421053 0.422441 0.420115 0.421782
#> Age.25-49    0.850877 0.851932 0.852029 0.851954
#> Age.50+      0.883436 0.882694 0.882849 0.882512
#> 
#> 
#> $percents$by_method
#> $percents$by_method$raw
#>                  Daily    Often     Rare    Never
#> Gender.Men   -1.120542 0.728261 0.553571 0.838710
#> Gender.Women -0.914440 0.620690 0.500000 0.793750
#> Age.15-24    -0.323013 0.568627 0.333333 0.421053
#> Age.25-49    -1.277271 0.773333 0.653061 0.850877
#> Age.50+      -1.049945 0.641509 0.525000 0.883436
#> 
#> $percents$by_method$expe
#>                  Daily    Often     Rare    Never
#> Gender.Men   -1.126346 0.728822 0.561332 0.836192
#> Gender.Women -0.910035 0.619934 0.493587 0.796514
#> Age.15-24    -0.321296 0.569154 0.329701 0.422441
#> Age.25-49    -1.279643 0.773990 0.653721 0.851932
#> Age.50+      -1.049973 0.639807 0.527472 0.882694
#> 
#> $percents$by_method$pure
#>                  Daily    Often     Rare    Never
#> Gender.Men   -1.134166 0.730833 0.563931 0.839402
#> Gender.Women -0.902259 0.617969 0.491210 0.793080
#> Age.15-24    -0.316503 0.567698 0.328690 0.420115
#> Age.25-49    -1.281633 0.775360 0.654244 0.852029
#> Age.50+      -1.049768 0.639537 0.527382 0.882849
#> 
#> $percents$by_method$expe.lin
#>                  Daily    Often     Rare    Never
#> Gender.Men   -1.126143 0.728896 0.561224 0.836023
#> Gender.Women -0.909877 0.620018 0.493507 0.796352
#> Age.15-24    -0.320395 0.569081 0.329532 0.421782
#> Age.25-49    -1.279823 0.774128 0.653741 0.851954
#> Age.50+      -1.049764 0.639949 0.527303 0.882512
#> 
#> 
#>