Computes statistics for a categorical supplementary variable

From MCA results, computes statistics (weights, coordinates, contributions, test-values, variances) for a categorical supplementary variable.

varsup(resmca, var)

Arguments

resmca
object of class 'MCA', 'speMCA', 'csMCA', 'stMCA' or 'multiMCA'

var
the categorical supplementary variable. It does not need to have been used at the MCA step.

Value

Returns a list:

weight

numeric vector of categories weights

coord

data frame of categories coordinates

cos2

data frame of categories square cosine

var

data frame of categories within variances, variance between and within categories and variable square correlation ratio (eta2)

v.test

data frame of categories test-values

References

Le Roux B. and Rouanet H., Multiple Correspondence Analysis, SAGE, Series: Quantitative Applications in the Social Sciences, Volume 163, CA:Thousand Oaks (2010).

Le Roux B. and Rouanet H., Geometric Data Analysis: From Correspondence Analysis to Stuctured Data Analysis, Kluwer Academic Publishers, Dordrecht (June 2004).

Author

Nicolas Robette

See also

Examples

## Performs a specific MCA on 'Music' example data set
## ignoring every 'NA' (i.e. 'not available') categories,
## and then computes statistics for age supplementary variable.
data(Music)
getindexcat(Music)

#>  [1] "FrenchPop.No"  "FrenchPop.Yes" "FrenchPop.NA"  "Rap.No"       
#>  [5] "Rap.Yes"       "Rap.NA"        "Rock.No"       "Rock.Yes"     
#>  [9] "Rock.NA"       "Jazz.No"       "Jazz.Yes"      "Jazz.NA"      
#> [13] "Classical.No"  "Classical.Yes" "Classical.NA"  "Gender.Men"   
#> [17] "Gender.Women"  "Age.15-24"     "Age.25-49"     "Age.50+"      
#> [21] "OnlyMus.Daily" "OnlyMus.Often" "OnlyMus.Rare"  "OnlyMus.Never"
#> [25] "Daily.No"      "Daily.Yes"    

mca <- speMCA(Music[,1:5],excl=c(3,6,9,12,15))
varsup(mca,Music$Age)

#> $weight
#> 15-24 25-49   50+ 
#>    76   201   223 
#> 
#> $coord
#>           dim.1     dim.2     dim.3     dim.4     dim.5
#> 15-24  0.090284  0.872172  0.218187 -0.108922 -0.107015
#> 25-49  0.001334  0.152716  0.199059  0.229128  0.049278
#> 50+   -0.031972 -0.434893 -0.253781 -0.169402 -0.007945
#> 
#> $cos2
#>          dim.1    dim.2    dim.3    dim.4    dim.5
#> 15-24 0.001461 0.136349 0.008533 0.002127 0.002053
#> 25-49 0.000001 0.015678 0.026637 0.035292 0.001632
#> 50+   0.000823 0.152261 0.051849 0.023103 0.000051
#> 
#> $var
#>            dim.1    dim.2    dim.3    dim.4    dim.5
#> 15-24   0.338684 0.285749 0.182635 0.363425 0.102039
#> 25-49   0.293869 0.202105 0.185404 0.193889 0.105982
#> 50+     0.262978 0.137471 0.171012 0.058844 0.146490
#> within  0.286904 0.185992 0.178564 0.159429 0.123450
#> between 0.000487 0.049248 0.009773 0.005904 0.000340
#> total   0.287391 0.235240 0.188337 0.165332 0.123789
#> eta2    0.001696 0.209352 0.051890 0.035707 0.002745
#> 
#> $v.test
#>           dim.1     dim.2     dim.3     dim.4     dim.5
#> 15-24  0.853861  8.248528  2.063497 -1.030127 -1.012093
#> 25-49  0.024436  2.797036  3.645817  4.196540  0.902533
#> 50+   -0.640818 -8.716556 -5.086531 -3.395328 -0.159234
#>